% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_utilities.R
\name{add_name_column}
\alias{add_name_column}
\title{Add a Name Column to a Data Frame}
\usage{
add_name_column(df, name)
}
\arguments{
\item{df}{A data frame to which the new column will be added.}

\item{name}{A character string that will be used to fill the new column.}
}
\value{
The modified data frame with an additional column named 'DataFrameName'
which contains the same value specified by the `name` parameter for all rows.
}
\description{
This function adds a new column to an existing data frame where all entries in
the new column are set to the specified name. This can be useful for tracking the
origin of data after merging several data frames.
}
\examples{
\dontrun{
  # Assuming 'your_dataframe' is your existing data frame and you want to
  # label it as "sample_data"
  labeled_df <- add_name_column(your_dataframe, "sample_data")
  print(labeled_df)
}

}
