% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_utilities.R
\name{addEntryToDF}
\alias{addEntryToDF}
\title{Append a new entry to a data frame}
\usage{
addEntryToDF(original_df, new_entry)
}
\arguments{
\item{original_df}{A data frame to which the new entry will be added.}

\item{new_entry}{A list or a data frame row that represents the new entry to be added.}
}
\value{
A data frame with the new entry appended to the original data frame.
}
\description{
This function takes an existing data frame and a new entry (as a row) and appends the new entry to the original data frame using `rbind`.
The resulting data frame, with the new entry added, is then returned.
}
\examples{
original_df <- data.frame(Name = c("Alice", "Bob"), Age = c(25, 30))
new_entry <- data.frame(Name = "Charlie", Age = 35)
addEntryToDF(original_df, new_entry)
}
