% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{addListNameToDFElements}
\alias{addListNameToDFElements}
\title{Add List Element Names to Data Frames Within a List}
\usage{
addListNameToDFElements(this_list, element_name)
}
\arguments{
\item{this_list}{A list of data frames.}

\item{element_name}{The name of the new column to be added to each data frame,
which will contain the name of the list element.}
}
\value{
The same list with each data frame now containing an additional column
        named as specified by `elementName`, filled with the name of the list element.
}
\description{
Iterates through a list of data frames and adds a new column to each data frame
with the name of the list element.
}
\examples{
# Assuming 'listOfDFs' is a list of data frames:
updatedList <- addListNameToDFElements(listOfDFs, "sourceName")
}
