% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intracellular_tailoring.R
\name{addRandomGRNs}
\alias{addRandomGRNs}
\title{Add Random Gene Regulatory Networks to a Cluster's GRN}
\usage{
addRandomGRNs(decipher_seurat_this_cluster, regulon_this_cluster_capped)
}
\arguments{
\item{decipher_seurat_this_cluster}{A Seurat object specific to a cluster, from which gene
names are extracted to generate random GRNs.}

\item{regulon_this_cluster_capped}{A data frame representing the capped GRN of the cluster.
This GRN is used as the reference to generate random GRNs.}
}
\value{
A data frame that combines the original capped GRN with additional random GRNs.
}
\description{
This function augments a gene regulatory network (GRN) for a specific cluster with random GRNs.
The random GRNs are generated based on a reference GRN to simulate additional regulatory
relationships. These random GRNs are combined with the original GRN to expand the dataset,
which might be useful for simulation or robustness testing.
}
\examples{
# Assuming 'decipher_seurat_cluster' is a Seurat object for a specific cluster and
# 'regulon_cluster_capped' is the existing GRN for that cluster:
enhanced_grn <- addRandomGRNs(
  decipher_seurat_this_cluster = decipher_seurat_cluster,
  regulon_this_cluster_capped = regulon_cluster_capped
)

}
