% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intracellular_scoring.R
\name{adjustScoresOrientation}
\alias{adjustScoresOrientation}
\title{Title}
\usage{
adjustScoresOrientation(scores, x, rotation, isTargetLabel)
}
\arguments{
\item{isTargetLabel}{}
}
\description{
Title
}
