% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_utilities.R
\name{assign_tf_family_colors_to_each_regulon}
\alias{assign_tf_family_colors_to_each_regulon}
\title{Assign Family Colors to Each Regulon}
\usage{
assign_tf_family_colors_to_each_regulon(top_30_regulons_delta_matrix)
}
\arguments{
\item{top_30_regulons_delta_matrix}{A matrix with regulons as rows. The names of the regulons must be set
as the row names of the matrix. It is assumed that the matrix contains the top 30 regulons based on
delta values.}
}
\value{
A named vector of colors corresponding to the transcription factors in the input matrix. The name of
        each element in the vector corresponds to a regulon, and the value is the color assigned based on
        the transcription factor family.
}
\description{
This function takes a matrix of the top 30 regulons and their delta values, and assigns a specific color
to each transcription factor (TF) based on its family. The five TF families included are NFKB, STAT, IRF,
AP1, and NFAT. Transcription factors that are not in these families are colored gray.
}
\examples{
# Assuming delta_matrix is a matrix with row names that include transcription factors
tf_colors <- assign_tf_family_colors_to_each_regulon(delta_matrix)
print(tf_colors)

}
