% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_data_wrangling.R
\name{calculate_p_value}
\alias{calculate_p_value}
\title{Calculate P-Value for a Given Value Against a Set of Base Values}
\usage{
calculate_p_value(base_values, real_value)
}
\arguments{
\item{base_values}{A numeric vector of base values assumed to be from a normal distribution.}

\item{real_value}{A single numeric value for which the p-value is to be calculated.}
}
\value{
A numeric value representing the two-sided p-value.
}
\description{
This function calculates the two-sided p-value of a given value based on its
standard score (z-score) in relation to a set of base values. It assumes that the
base values are normally distributed. The function computes the mean and standard
deviation of the base values, calculates the z-score of the given value, and then
derives the p-value using the standard normal distribution.
}
\examples{
base_values <- rnorm(100, mean = 50, sd = 10)
real_value <- 60
p_value <- calculate_p_value(base_values, real_value)
print(p_value)

}
