% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_meta_cells.R
\name{calculate_suggested_number_of_metacell_neighbours}
\alias{calculate_suggested_number_of_metacell_neighbours}
\title{Calculate Suggested Number of Metacell Neighbours}
\usage{
calculate_suggested_number_of_metacell_neighbours(
  seurat_object,
  param_min_n_cells
)
}
\arguments{
\item{seurat_object}{A Seurat object containing single-cell RNA-seq data.}

\item{param_min_n_cells}{The minimum number of cells required per cluster-condition combination
after adjustment for the median UMI count.}
}
\value{
An integer representing the suggested number of neighbours (k) to ensure that all
cluster-condition combinations have a sufficient number of cells based on UMI count criteria.
}
\description{
This function estimates the suggested number of metacell neighbours for a Seurat object. It
iteratively calculates the median UMI counts per cluster and condition, adjusting the counts
to find a minimum k (number of neighbours) such that when divided by k, the count for each
cluster-condition combination remains above a specified minimum threshold. The process involves
filtering out cluster-condition combinations that fall below this threshold until a stable k is found.
}
\examples{
# Assuming 'seurat_object' is already created and initialized with necessary data
min_k <- calculate_suggested_number_of_metacell_neighbours(seurat_object, 50)

}
