% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_meta_cells.R
\name{calculateDistVarF}
\alias{calculateDistVarF}
\title{Calculate Distance Matrix from Variable Features of a Seurat Object}
\usage{
calculateDistVarF(seuratObject)
}
\arguments{
\item{seuratObject}{A Seurat object containing single-cell RNA sequencing data.}
}
\value{
A distance matrix computed from the variable features of the RNA assay
in the Seurat object.
}
\description{
This function calculates a distance matrix based on the variable features
of a Seurat object. It identifies variable features in the RNA assay of the
Seurat object and computes a distance matrix using these features.
}
\examples{
# seuratObject is a pre-existing Seurat object
distanceMatrix <- calculateDistVarF(seuratObject)
}
