% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{calculateEnrichmentStatistics}
\alias{calculateEnrichmentStatistics}
\title{Calculate Enrichment Statistics}
\usage{
calculateEnrichmentStatistics(other_gene_set, my_gene_set, base_num)
}
\arguments{
\item{other_gene_set}{A vector of gene symbols representing the comparison gene set.}

\item{my_gene_set}{A vector of gene symbols representing the gene set of interest.}

\item{base_num}{A numeric value representing the total number of possible genes.}
}
\value{
A list containing the following components:
\itemize{
  \item{p_value}{The p-value calculated using the hypergeometric test.}
  \item{genes}{A semicolon-separated string of genes found in both gene sets.}
  \item{gene_in_my_set_in_other_set}{The number of genes in the gene set of interest that are also in the comparison set.}
  \item{gene_in_other_set}{The total number of genes in the comparison set.}
  \item{gene_in_my_set}{The total number of genes in the gene set of interest.}
}
}
\description{
This function calculates enrichment statistics, including a p-value, for a specified gene set
against another gene set. It uses the hypergeometric test to compute the p-value, which indicates
the probability of observing the given overlap between the two gene sets.
}
\examples{
# Example usage:
my_genes <- c("gene1", "gene2", "gene3")
other_genes <- c("gene2", "gene4", "gene5")
result <- calculateEnrichmentStatistics(other_genes, my_genes, 1000)

}
