% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercellular_scoring.R
\name{calculateInteractionDeltas}
\alias{calculateInteractionDeltas}
\title{Calculate Interaction Deltas for Cluster-Specific Interaction Potentials}
\usage{
calculateInteractionDeltas(
  interaction_potentials_matrix_this_cluster,
  decipher_seurat_lr
)
}
\arguments{
\item{interaction_potentials_matrix_this_cluster}{A matrix of interaction potentials for a specific cluster.}

\item{decipher_seurat_lr}{A Seurat object from which metadata will be used to annotate the newly created Seurat object.}
}
\value{
A data frame containing the interaction deltas, including log-fold changes, p-values,
        adjusted p-values, and gene names, filtered by significance based on adjusted p-value.
}
\description{
This function creates a new Seurat object using a matrix of interaction potentials specific to a cluster.
It then performs differential expression analysis to calculate interaction deltas between conditions
specified in the metadata. This analysis helps identify significant interactions in terms of adjusted p-values.
}
\examples{
# Assuming 'interaction_matrix' is a matrix and 'decipher_seurat' is a Seurat object:
interaction_deltas <- calculateInteractionDeltas(
  interaction_potentials_matrix_this_cluster = interaction_matrix,
  decipher_seurat_lr = decipher_seurat
)

}
