% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercellular_scoring.R
\name{calculateInteractionDeltasAllClusters}
\alias{calculateInteractionDeltasAllClusters}
\title{Calculate Interaction Deltas for All Clusters}
\usage{
calculateInteractionDeltasAllClusters(
  interaction_potentials_matrix_all_clusters,
  decipher_seurat_lr
)
}
\arguments{
\item{interaction_potentials_matrix_all_clusters}{A list of interaction potentials matrices for each cluster, typically obtained from `getInteractionPotentialsMatrixAllClusters`.}

\item{decipher_seurat_lr}{A Seurat object containing ligand-receptor interaction data.}
}
\value{
A list where each element corresponds to a cluster and contains the interaction deltas for that cluster.
}
\description{
This function calculates the interaction deltas for each cluster based on the interaction potentials matrix.
}
\details{
The function iterates through each cluster in the `interaction_potentials_matrix_all_clusters` list, calculates the interaction deltas using the `calculateInteractionDeltas` function, and stores the results in a list with each element corresponding to a cluster.
}
\examples{
\dontrun{
interaction_potentials_matrix_all_clusters <- getInteractionPotentialsMatrixAllClusters(
  decipher_seurat, decipher_seurat_this_cluster, L_set_relevant_features, TRUE)
decipher_seurat_lr <- CreateSeuratObject(counts = your_ligand_receptor_counts_matrix)
interaction_deltas <- calculateInteractionDeltasAllClusters(interaction_potentials_matrix_all_clusters, decipher_seurat_lr)
}

}
