% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercellular_scoring.R
\name{calculateInteractionMatrix}
\alias{calculateInteractionMatrix}
\title{Calculate Interaction Potential Matrix}
\usage{
calculateInteractionMatrix(receptorMatrix, conditionVector, ligandMeans, LRSet)
}
\arguments{
\item{receptorMatrix}{A matrix with receptors as row names and samples as columns.}

\item{conditionVector}{A vector indicating the condition (e.g., 'case' or 'control')
for each sample in `receptorMatrix`.}

\item{ligandMeans}{A data frame with mean values of ligands under different conditions.}

\item{LRSet}{A data frame representing ligand-receptor pairs and their interactions.}
}
\value{
An interaction potential matrix where rows represent ligand-receptor interactions
and columns represent samples. Each element in the matrix is the calculated interaction
potential for a ligand-receptor pair under a specific condition.
}
\description{
This function computes an interaction potential matrix for ligand-receptor pairs
based on the receptor expression matrix, a condition vector, mean ligand values,
and a ligand-receptor set (LRSet). It calculates the interaction potential for each
ligand-receptor pair under different conditions (e.g., 'case' and 'control').
}
\examples{
# receptorMatrix is a matrix with receptor genes as row names
# conditionVector is a vector indicating the condition of each sample
# ligandMeans is a data frame with mean ligand values under different conditions
# LRSet is a data frame with ligand-receptor pairs
interactionMatrix <- calculateInteractionMatrix(receptorMatrix, conditionVector, ligandMeans, LRSet)
}
