% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_meta_cells.R
\name{calculateMinimumN}
\alias{calculateMinimumN}
\title{calculateMinimumN}
\usage{
calculateMinimumN(groups, min_counts, paramPairings)
}
\arguments{
\item{groups}{groups from createGroupsFromPairings}

\item{min_counts}{the minimum number of counts accepted per cluster/condition}

\item{paramPairings}{a data frame of comparisons (case and control columns) whose values correspond to clusters}
}
\value{
minimum size of each comparison
}
\description{
calculateMinimumN
}
