% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_meta_cells.R
\name{calculatePseudoBulkCell}
\alias{calculatePseudoBulkCell}
\title{Calculate Pseudo Bulk Cell Data}
\usage{
calculatePseudoBulkCell(seuratObject, numNearestNeighbors, numMetaCells)
}
\arguments{
\item{seuratObject}{A Seurat object containing single-cell expression data.}

\item{numNearestNeighbors}{Integer, the number of nearest neighbors to consider.}

\item{numMetaCells}{Integer, the number of meta cells to calculate.}
}
\value{
A matrix representing pseudo bulk data calculated from the Seurat object.
}
\description{
This function calculates pseudo bulk cell data from a Seurat object.
It first retrieves the RNA counts matrix, then computes a distance matrix,
and finally calculates the pseudo bulk matrix using the specified number
of nearest neighbors and meta cells.
}
\examples{
# seuratObject is a pre-existing Seurat object
pseudoBulkMatrix <- calculatePseudoBulkCell(seuratObject, 5, 10)
}
