% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_scoring.R
\name{calculateScoresByCluster}
\alias{calculateScoresByCluster}
\title{Calculate Mean Decipher Scores by Cluster}
\usage{
calculateScoresByCluster(scores_by_regulon_cluster)
}
\arguments{
\item{scoresByRegulonCluster}{A data frame containing 'interaction', 'decipher_score',
'ligand', and 'receptor' columns. It represents scores by regulon and cluster.}
}
\value{
A data frame with the mean decipher score for each interaction,
        sorted in descending order of scores. The returned data frame includes
        'interaction', 'decipher_score', 'ligand', 'receptor', and a new column
        'sender_cluster' with a fixed value 'mixed'.
}
\description{
This function calculates the mean decipher score for each interaction within a cluster.
It also joins additional information (ligand and receptor) to the resulting data frame.
}
