% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intracellular_tailoring.R
\name{capRegulon}
\alias{capRegulon}
\title{Cap Regulon by Top N Targets}
\usage{
capRegulon(this_grn, n_top = 100)
}
\arguments{
\item{this_grn}{A data frame representing the gene regulatory network, expected to contain
a column 'source'.}

\item{n_top}{The maximum number of top entries to retain for each source.
Default is 100.}
}
\value{
A data frame similar to `this_grn` but with each source limited to a maximum of `n_top` entries.
}
\description{
This function limits the number of targets for each source in a gene regulatory network (GRN)
to the top N entries.
}
\examples{
# Assuming 'grn' is a gene regulatory network data frame:
cappedRegulon <- capRegulon(grn, n_top = 100)
}
