% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intracellular_tailoring.R
\name{capRegulonsAllClusters}
\alias{capRegulonsAllClusters}
\title{Cap Regulons for All Clusters}
\usage{
capRegulonsAllClusters(
  regulons_all_clusters,
  decipher_seurat,
  flag.normalize.non.log
)
}
\arguments{
\item{regulons_all_clusters}{A list of regulons for each cluster, typically obtained from `getRegulonsAllClusters`.}

\item{decipher_seurat}{A Seurat object containing single-cell RNA-seq data with cluster and condition metadata.}

\item{flag.normalize.non.log}{A logical flag indicating whether to normalize non-log-transformed data.}
}
\value{
A list where each element corresponds to a cluster and contains the capped regulons for that cluster.
}
\description{
This function processes and caps the regulons for each cluster, normalizing the Seurat object if specified.
}
\details{
The function iterates through each unique cluster in the `decipher_seurat` object, subsets the Seurat object for the cluster, normalizes the data if `flag.normalize.non.log` is TRUE, caps the regulons to the top 40 features, and adds random gene regulatory networks (GRNs) to the capped regulons.
}
\examples{
\dontrun{
regulons_all_clusters <- getRegulonsAllClusters(filepath, seurat_object)
decipher_seurat <- CreateSeuratObject(counts = your_counts_matrix)
capped_regulons <- capRegulonsAllClusters(regulons_all_clusters, decipher_seurat, TRUE)
}

}
