% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_utilities.R
\name{checkAndAdaptDecipherVersion}
\alias{checkAndAdaptDecipherVersion}
\title{Check and Adapt Decipher Dataframe Format}
\usage{
checkAndAdaptDecipherVersion(regulon_deltas_by_cluster)
}
\arguments{
\item{regulon_deltas_by_cluster}{A list of dataframes, each representing regulon
delta values by cluster. Expected to possibly contain a column named "delta.pagoda",
which will be renamed to "deltaPagoda".}
}
\value{
The same list of dataframes passed as input, potentially with column names
updated from "delta.pagoda" to "deltaPagoda".
}
\description{
Examines the first dataframe within a list of dataframes to determine if it uses an
older version of Decipher naming conventions (specifically, checks for the presence
of "delta.pagoda" column name). If this older convention is detected, the function
updates the column name to the newer version ("deltaPagoda") across all dataframes in
the list. It prints a message indicating whether the data was updated or is already
in the expected format.
}
\examples{
\dontrun{
  # Assuming 'regulon_deltas' is your list of dataframes with regulon delta values
  updated_deltas <- checkAndAdaptDecipherVersion(regulon_deltas)
}

}
