% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_utilities.R
\name{clean_names}
\alias{clean_names}
\title{Clean special characters in string names}
\usage{
clean_names(this_string)
}
\arguments{
\item{this_string}{A single string or character vector where patterns need to be replaced.}
}
\value{
A string or character vector with the specified patterns replaced by conventional characters.
}
\description{
This function modifies strings by replacing specific patterns with more conventional characters.
It replaces '_minus_' with '-' and '_plus_' with '+'. This can be particularly useful in cleaning up variable names or labels that have been encoded with special characters to adhere to programming language syntax constraints.
}
\examples{
clean_names("rate_minus_inflation_plus_growth")
}
