% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_utilities.R
\name{convert_text_patterns}
\alias{convert_text_patterns}
\title{Convert Specific Text Patterns in Strings}
\usage{
convert_text_patterns(text_vec)
}
\arguments{
\item{text_vec}{A character vector containing the text to be processed.}
}
\value{
A character vector with the specified text patterns replaced.
}
\description{
This function takes a vector of strings and replaces specific text patterns
according to the following rules:
- Replaces "_plus_" with "+"
- Replaces "_minus_" with "-"
- Replaces "monocytes" with " Mono"
}
\examples{
text_vec <- c("T_cells_plus_", "B_cells_minus_", "monocytes")
converted_text <- convert_text_patterns(text_vec)
print(converted_text)

}
