% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convertHumanSymbolsToMouse}
\alias{convertHumanSymbolsToMouse}
\title{Convert Human Gene Symbols to Mouse Orthologs}
\usage{
convertHumanSymbolsToMouse(human_symbols)
}
\arguments{
\item{human_symbols}{A vector of human gene symbols to be converted.}
}
\value{
A vector of mouse gene symbols corresponding to the human input.
        NA values and duplicates are removed from the output.
}
\description{
This function takes a vector of human gene symbols and finds their corresponding
mouse orthologs using the babelgene package. It returns a vector of mouse gene
symbols that are orthologs of the given human genes.
}
\examples{
# Example usage:
human_genes <- c("BRCA1", "TP53")
mouse_genes <- convertHumanSymbolsToMouse(human_genes)

}
