% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convertListOfMatricesToMatrix}
\alias{convertListOfMatricesToMatrix}
\title{Convert a List of Matrices to a Single Matrix}
\usage{
convertListOfMatricesToMatrix(listOfMatrices)
}
\arguments{
\item{listOfMatrices}{A list where each element is a matrix of potentially varying number
of rows but the same number of columns.}
}
\value{
A single matrix composed of all matrices from the list combined row-wise. The
        function assumes that all matrices in the list have the same number of columns.
}
\description{
This function concatenates a list of matrices into a single matrix by binding them row-wise.
It iterates through each matrix in the list, starting with the first matrix, and subsequently
binds each following matrix to the result of the previous concatenations. This is useful for
combining data from similar matrices stored in a list into a single matrix structure.
}
\examples{
# Assuming 'mat_list' is a list containing several matrices:
combined_matrix <- convertListOfMatricesToMatrix(mat_list)

}
