% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convertLsetToMouse}
\alias{convertLsetToMouse}
\title{Convert Human Ligand-Receptor Set to Mouse}
\usage{
convertLsetToMouse(human_l_set)
}
\arguments{
\item{human_l_set}{A data frame representing the human ligand-receptor set.
This data frame must contain the columns 'ligand' and 'receptor'.}
}
\value{
A data frame representing the mouse ligand-receptor set.
        The output includes 'ligand' and 'receptor' columns for mouse genes,
        along with 'old_ligand_name' and 'old_receptor_name' columns to reference
        the original human genes. It also includes an 'interaction' column
        representing the ligand-receptor pairs.
}
\description{
This function converts a human ligand-receptor set to its mouse equivalent.
It utilizes ortholog mapping to find corresponding mouse genes for human ligands and receptors.
The function requires an input data frame with 'ligand' and 'receptor' columns.
}
\examples{
# Example usage:
human_l_set <- data.frame(
  ligand = c("LIG1", "LIG2"),
  receptor = c("REC1", "REC2")
)
mouse_l_set <- convertLsetToMouse(human_l_set)

}
