% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_meta_cells.R
\name{CorDist}
\alias{CorDist}
\title{Calculate Distance Matrix Using Correlation}
\usage{
CorDist(matrix, pcaFeats = NULL, corMethod = "spearman")
}
\arguments{
\item{matrix}{A numeric matrix on which the distance calculation is based.}

\item{pcaFeats}{An optional integer specifying the number of principal components
to retain for distance calculation. If NULL, PCA is not performed.}

\item{corMethod}{A character string indicating the correlation method to be used.
Can be one of 'pearson', 'kendall', or 'spearman'.}
}
\value{
A distance matrix, with each element representing the distance
calculated based on the correlation between columns of the input matrix.
}
\description{
This function computes a distance matrix based on the correlation of the input matrix.
Optionally, it can first perform a principal component analysis (PCA) and use the
specified number of PCA features for the distance calculation.
}
\examples{
dataMatrix <- matrix(rnorm(100), ncol = 10)
distanceMatrix <- CorDist(dataMatrix)
# With PCA transformation
pcaDistanceMatrix <- CorDist(dataMatrix, pcaFeats = 5)
}
