% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intracellular_scoring.R
\name{createAndPreparePagoda2}
\alias{createAndPreparePagoda2}
\title{Create and prepare pagoda2}
\usage{
createAndPreparePagoda2(counts)
}
\arguments{
\item{counts}{}
}
\description{
Create and prepare pagoda2
}
