% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_meta_cells.R
\name{createGroupsFromPairings}
\alias{createGroupsFromPairings}
\title{createGroupsFromPairings}
\usage{
createGroupsFromPairings(paramPairings)
}
\arguments{
\item{paramPairings}{a data frame of comparisons (case and control columns) whose values correspond to clusters}
}
\value{
a group from which to calculate minimum sizes
}
\description{
createGroupsFromPairings
}
