% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_data_wrangling.R
\name{do_t_test_by_feature_by_grouping_factor}
\alias{do_t_test_by_feature_by_grouping_factor}
\title{Perform t-tests by feature, grouped by a factor}
\usage{
do_t_test_by_feature_by_grouping_factor(data_matrix, group_factor)
}
\arguments{
\item{data_matrix}{A matrix where rows represent samples and columns represent features.}

\item{group_factor}{A factor specifying the grouping of samples.}
}
\value{
A named vector of p-values, where each p-value corresponds to a feature in the data matrix.
}
\description{
This function conducts t-tests for each feature in a data matrix, comparing values across groups defined by a grouping factor.
}
\details{
This function iterates over each feature in the data matrix, conducts a t-test comparing the values across groups defined by the grouping factor, and returns the p-values for each feature.
}
\examples{
data_matrix <- matrix(rnorm(100), nrow = 10)
group_factor <- factor(rep(c("A", "B"), each = 5))
do_t_test_by_feature_by_grouping_factor(data_matrix, group_factor)

}
