% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrichResults}
\alias{enrichResults}
\title{Enrichment Analysis of Differential Expression and Regulon Results}
\usage{
enrichResults(
  de_markers_this_cluster,
  significant_regulon_deltas_this_cluster,
  regulon_this_cluster,
  enrichr_database
)
}
\arguments{
\item{de_markers_this_cluster}{A dataframe containing differential expression markers with
columns for average log2 fold change ('avg_log2FC') and gene names ('gene').}

\item{significant_regulon_deltas_this_cluster}{A dataframe containing significant regulons with
at least a column for regulon names ('name').}

\item{regulon_this_cluster}{A dataframe listing target genes for each regulon source.}

\item{enrichr_database}{A list of databases, each an enrichment dataset to be used for analysis.}
}
\value{
A list where each element corresponds to a regulon and contains a dataframe of combined
        enrichment results across databases, including adjusted p-values and associated terms.
}
\description{
This function performs enrichment analysis using predefined gene sets from differential
expression markers and regulon analysis. It separates positive and negative markers, identifies
top markers based on log fold changes, and performs enrichment using a given Enrichr database.
The function also handles multiple enrichment databases and adjusts p-values for multiple testing.
}
\examples{
# Assuming appropriate data structures are already created:
results <- enrichResults(
  de_markers_this_cluster = differential_markers,
  significant_regulon_deltas_this_cluster = significant_regulons,
  regulon_this_cluster = regulon_targets,
  enrichr_database = loaded_enrichr_dbs
)

}
