% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrichResultsAllClusters}
\alias{enrichResultsAllClusters}
\title{Enrich Results for All Clusters}
\usage{
enrichResultsAllClusters(
  de_markers_all_clusters,
  significant_regulon_deltas_all_clusters,
  regulons_all_clusters,
  enrichr_database
)
}
\arguments{
\item{de_markers_all_clusters}{A list of differentially expressed markers for each cluster, typically obtained from `FindMarkersAllClusters`.}

\item{significant_regulon_deltas_all_clusters}{A list of significant regulon deltas for each cluster.}

\item{regulons_all_clusters}{A list of regulons for each cluster.}

\item{enrichr_database}{A character vector specifying the database to use for enrichment analysis.}
}
\value{
A list where each element corresponds to a cluster and contains the enrichment results for that cluster.
}
\description{
This function performs enrichment analysis on differentially expressed markers for each cluster using a specified database.
}
\details{
The function iterates through each cluster in the `de_markers_all_clusters` list, performs enrichment analysis on the differentially expressed markers using the `enrichResults` function, and stores the results in a list. The enrichment analysis uses the specified `enrichr_database`.
}
\examples{
\dontrun{
de_markers_all_clusters <- FindMarkersAllClusters(decipher_seurat, TRUE)
significant_regulon_deltas_all_clusters <- getSignificantRegulonsAllClusters(regulon_deltas_all_clusters)
regulons_all_clusters <- getRegulonsAllClusters(filepath, decipher_seurat)
enrichr_database <- "GO_Biological_Process_2021"
enriched_results_all_clusters <- enrichResultsAllClusters(
  de_markers_all_clusters, significant_regulon_deltas_all_clusters, regulons_all_clusters, enrichr_database)
}

}
