% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decipher_scoring.R
\name{extractDecipherResults}
\alias{extractDecipherResults}
\title{Extract and Analyze Decipher Results from Random Forest and Interaction Data}
\usage{
extractDecipherResults(
  random_forest_results,
  interaction_potentials_matrix_clusters,
  data_this_cluster_downsampled_receptors,
  selected_lr_pairs,
  this.tf,
  val.this.tf,
  tf.merged
)
}
\arguments{
\item{random_forest_results}{A randomForest object containing the results from a random forest analysis.}

\item{interaction_potentials_matrix_clusters}{A matrix of interaction potentials between clusters.}

\item{data_this_cluster_downsampled_receptors}{A matrix or dataframe containing receptor data for a downsampled cluster.}

\item{selected_lr_pairs}{A dataframe specifying ligand-receptor pairs.}

\item{this_tf}{The specific transcription factor (TF) being analyzed.}

\item{val.this_tf}{The value associated with the transcription factor in the analysis.}
}
\value{
A dataframe containing a comprehensive summary of interactions, importance metrics,
        correlation scores, and additional computed values related to the transcription factor.
}
\description{
This function processes the results from a random forest analysis, computing the importance
of features and their correlations with transcription factors (TFs). It merges these results
with interaction potential data to create a detailed summary of interactions, including ligands,
receptors, and their importance metrics. Additionally, it evaluates correlations using both
Spearman and Pearson methods and calculates contributions based on these correlations.
}
\examples{
# Assuming you have all the necessary data prepared:
decipher_results_df <- extractDecipherResults(
  random_forest_results = rf_results,
  interaction_potentials_matrix_clusters = interaction_matrix,
  data_this_cluster_downsampled_receptors = receptor_data,
  selected_lr_pairs = lr_pairs,
  this_tf = "TF_name",
  val.this_tf = 0.5
)

}
