% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_utilities.R
\name{filter_threshold}
\alias{filter_threshold}
\title{Filter Data Based on Threshold}
\usage{
filter_threshold(data, threshold_main, threshold_fallback)
}
\arguments{
\item{data}{A data frame containing the dataset to be filtered.}

\item{threshold_main}{The main threshold value used for filtering.}

\item{threshold_fallback}{The fallback threshold value used if no rows meet the main threshold.}
}
\value{
A subset of the input data frame containing rows that meet either the main threshold or the fallback threshold.
}
\description{
This function filters a dataset based on two thresholds. It selects rows that meet the main threshold, and if none do, it selects rows with the fallback threshold.
}
\examples{
my_data <- data.frame(
  ID = c(1, 2, 3, 4, 5),
  Value = c(10, 20, 30, 40, 50),
  Threshold = c(15, 25, 35, 45, 55)
)
filter_threshold(my_data, 30, 25)

}
