% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercellular_scoring.R
\name{filterIntPotByDeltas}
\alias{filterIntPotByDeltas}
\title{Filter Interaction Potentials by Deltas}
\usage{
filterIntPotByDeltas(
  interaction_potentials_matrix_all_clusters,
  interaction_deltas_all_clusters
)
}
\arguments{
\item{interaction_potentials_matrix_all_clusters}{A list of interaction potentials matrices for each cluster, typically obtained from `getInteractionPotentialsMatrixAllClusters`.}

\item{interaction_deltas_all_clusters}{A list of interaction deltas for each cluster, typically obtained from `calculateInteractionDeltasAllClusters`.}
}
\value{
A list where each element corresponds to a cluster and contains the filtered interaction potentials matrix for that cluster.
}
\description{
This function filters the interaction potentials matrix for each cluster based on the interaction deltas.
}
\details{
The function iterates through each cluster in the `interaction_potentials_matrix_all_clusters` list, filters the interaction potentials matrix based on the interaction deltas, and stores the results in a list with each element corresponding to a cluster.
}
\examples{
\dontrun{
interaction_potentials_matrix_all_clusters <- getInteractionPotentialsMatrixAllClusters(
  decipher_seurat, decipher_seurat_this_cluster, L_set_relevant_features, TRUE)
interaction_deltas_all_clusters <- calculateInteractionDeltasAllClusters(interaction_potentials_matrix_all_clusters, decipher_seurat_lr)
filtered_interaction_potentials_matrix <- filterIntPotByDeltas(interaction_potentials_matrix_all_clusters, interaction_deltas_all_clusters)
}

}
