% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decipher_scoring.R
\name{FindLRMarkersAllClustersWParamPairings}
\alias{FindLRMarkersAllClustersWParamPairings}
\title{Find LR Marker Genes for Ligand-Receptor Pairs in All Clusters}
\usage{
FindLRMarkersAllClustersWParamPairings(
  decipher_seurat,
  rf_results_all_clusters,
  flag.normalize.non.log,
  paramPairings,
  random.seed
)
}
\arguments{
\item{decipher_seurat}{A Seurat object containing single-cell RNA-seq data with cluster and condition metadata.}

\item{rf_results_all_clusters}{A list of random forest results for each cluster.}

\item{flag.normalize.non.log}{A logical flag indicating whether to normalize non-log-transformed data.}

\item{paramPairings}{}
}
\value{
A list where each element corresponds to a cluster and contains the marker genes for that cluster.
}
\description{
This function identifies marker genes for ligand-receptor pairs in each cluster of a Seurat object based on random forest results.
}
\details{
The function iterates through each unique cluster in the `decipher_seurat` object, subsets the Seurat object for the cluster, normalizes the data if necessary, and identifies marker genes for ligand-receptor pairs using the `FindMarkers` function.
}
\examples{
\dontrun{
decipher_seurat <- CreateSeuratObject(counts = your_counts_matrix)
rf_results_all_clusters <- getRandomForestWeightsAllClusters(
  decipher_seurat, significant_regulon_deltas_all_clusters, regulon_scores_all_clusters,
  interaction_potentials_matrix_clusters_all_clusters, L_set_relevant_features_all_clusters, TRUE)
lr_markers_all_clusters <- FindLRMarkersAllClusters(decipher_seurat, rf_results_all_clusters, TRUE)
}

}
