% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decipher_scoring.R
\name{FindMarkersAllClusters}
\alias{FindMarkersAllClusters}
\title{Find Differentially Expressed Markers for All Clusters}
\usage{
FindMarkersAllClusters(decipher_seurat, flag.normalize.non.log)
}
\arguments{
\item{decipher_seurat}{A Seurat object containing single-cell RNA-seq data with cluster and condition metadata.}

\item{flag.normalize.non.log}{A logical flag indicating whether to normalize non-log-transformed data.}
}
\value{
A list where each element corresponds to a cluster and contains the differentially expressed markers for that cluster.
}
\description{
This function identifies differentially expressed markers for each cluster in a Seurat object.
}
\details{
The function iterates through each unique cluster in the `decipher_seurat` object, subsets the Seurat object for the cluster, normalizes the data if necessary, and identifies differentially expressed markers using the `FindMarkers` function. The results for each cluster are stored in a list.
}
\examples{
\dontrun{
decipher_seurat <- CreateSeuratObject(counts = your_counts_matrix)
markers_all_clusters <- FindMarkersAllClusters(decipher_seurat, TRUE)
}

}
