% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_data_wrangling.R
\name{generateComparisonObjectFromNicheNet}
\alias{generateComparisonObjectFromNicheNet}
\title{Generate Comparison Object from NicheNet Analysis}
\usage{
generateComparisonObjectFromNicheNet(nichenetObject)
}
\arguments{
\item{nichenetObject}{A list containing NicheNet analysis results, where each
element is a tibble with at least the columns `test_ligand` and `aupr_corrected`.}
}
\value{
A modified version of the input list where each tibble has been processed
to include 'ligand', 'score', 'receiver_cluster', 'receptor', 'interaction', and
'sender_cluster' columns, with 'receptor' and 'sender_cluster' set to default values.
}
\description{
This function processes the results from a NicheNet analysis, stored within
a list of tibbles, to format them for comparison purposes. It renames columns
to standardize ligand and score nomenclature, and adds new columns for receiver
cluster, receptor (set to "unknown"), interaction (combining ligand and receptor),
and sender cluster (set to "mixed").
}
\examples{
\dontrun{
  # Assuming 'nichenet_results' is your list of NicheNet analysis results
  comparison_object <- generateComparisonObjectFromNicheNet(nichenet_results)
  print(comparison_object)
}

}
