% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_meta_cells.R
\name{generateMetaCellMatrices}
\alias{generateMetaCellMatrices}
\title{Generate Pseudobulk Matrices from Seurat Object}
\usage{
generateMetaCellMatrices(
  seuratObj,
  paramMinMetaCells = 100,
  paramMaxMetaCells = 600,
  paramMaxScCells,
  paramK
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{paramMinMetaCells}{The minimum number of meta cells to generate.}

\item{paramMaxMetaCells}{The maximum number of meta cells to generate.}

\item{paramMaxScCells}{The maximum number of single cells to consider within each condition.}

\item{paramK}{Parameter 'k' used in calculatePseudoBulkCell function.}
}
\value{
A list of pseudobulk matrices organized by cluster and condition.
}
\description{
Iterates over clusters and conditions in a Seurat object to generate pseudobulk matrices.
It handles minimum and maximum cell count constraints for pseudobulk generation.
}
