% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_meta_cells.R
\name{generatePseudoBulkSeurat}
\alias{generatePseudoBulkSeurat}
\title{Generate PseudoBulk Seurat Object}
\usage{
generatePseudoBulkSeurat(pseudobulkList)
}
\arguments{
\item{pseudobulkList}{A list of matrices representing pseudo bulk data,
named by clusters and conditions.}
}
\value{
A Seurat object created from the combined pseudo bulk data.
}
\description{
Joins pseudo bulk matrices and creates a Seurat object with the combined data
and corresponding metadata.
}
