% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intracellular_scoring.R
\name{generateRandomGRNsFromReference}
\alias{generateRandomGRNsFromReference}
\title{Generate Random GRNs From Reference}
\usage{
generateRandomGRNsFromReference(all_genes, reference_grns)
}
\arguments{
\item{all_genes}{A vector of all possible genes to be sampled.}

\item{reference_grns}{A data frame representing the reference GRNs,
containing at least the 'source' column.}
}
\value{
A data frame representing the randomly generated GRNs,
        with each row corresponding to a gene from the sampled genes,
        and columns for source, target, coef_mean, coef_abs, p, and X.logp.
}
\description{
This function generates random Gene Regulatory Networks (GRNs)
based on a provided reference set of GRNs. For each unique source
in the reference, it samples a specified number of genes and
constructs a data frame with default statistical values.
}
\examples{
# Example usage:
all_genes_example <- c("gene1", "gene2", "gene3", "gene4")
reference_grns_example <- data.frame(
  source = c("A", "B", "A", "C"),
  other_columns = 1:4
)
random_grns <- generateRandomGRNsFromReference(all_genes_example, reference_grns_example)

}
