% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dataset.R
\name{generateSampleSeuratFromExperimentHub}
\alias{generateSampleSeuratFromExperimentHub}
\title{Generate a Sample Seurat Object from ExperimentHub}
\usage{
generateSampleSeuratFromExperimentHub(min_cells_per_cluster_condition)
}
\arguments{
\item{min_cells_per_cluster_condition}{The minimum number of cells required per cluster-condition combination.}
}
\value{
A preprocessed Seurat object with clusters filtered to meet the cell count criteria.
}
\description{
This function loads a dataset from ExperimentHub, specifically searching for a dataset related to "Kang".
It processes the data, converts it to a Seurat object, and applies various preprocessing steps, including
normalization and filtering based on specified metadata criteria. It also ensures that clusters meet a
specified minimum number of cells per condition.
}
\examples{
# Generate a Seurat object with a minimum of 100 cells per cluster-condition combination
seurat_object <- generateSampleSeuratFromExperimentHub(min_cells_per_cluster_condition = 100)

}
