% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_utilities.R
\name{get_abs_max}
\alias{get_abs_max}
\title{Calculate Absolute Maximum Value}
\usage{
get_abs_max(row)
}
\arguments{
\item{row}{A numeric vector for which the absolute maximum value is sought.}
}
\value{
A single numeric value representing the maximum absolute value
found in the input vector.
}
\description{
Computes the maximum value in absolute terms from a numeric vector. This
function is useful for identifying the largest magnitude value, regardless
of its sign, within a given set of numbers.
}
\examples{
\dontrun{
  sample_vector <- c(-10, 2, 3, -4, 5)
  max_val <- get_abs_max(sample_vector)
  print(max_val) # Outputs: 10
}

}
