% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_data_wrangling.R
\name{get_top_regulons}
\alias{get_top_regulons}
\title{Get Top Regulons Based on Delta Threshold}
\usage{
get_top_regulons(df, delta_threshold)
}
\arguments{
\item{df}{A dataframe containing regulon data, expected to include columns
`deltaPagoda` for regulon scores and `name` for regulon names.}

\item{delta_threshold}{A numeric value specifying the minimum absolute value of
`deltaPagoda` required for a regulon to be considered as "top" or significant.}
}
\value{
A vector of unique regulon names that meet the specified delta threshold,
ordered by their `deltaPagoda` scores in descending order.
}
\description{
Filters and identifies the top regulons from a dataframe based on a specified
delta threshold. Regulons with their `deltaPagoda` absolute values exceeding this
threshold are selected and ordered by `deltaPagoda` in descending order. This function
is useful for identifying the most significant regulons by their changes in Pagoda
scores, possibly indicating a significant regulatory effect or change.
}
\examples{
\dontrun{
  # Assuming 'regulon_df' is your dataframe and you're interested in regulons
  # with a deltaPagoda score above 0.5
  top_regulons <- get_top_regulons(regulon_df, 0.5)
  print(top_regulons)
}

}
