% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{getDifferentiallyExpressedTargetsForRegulons}
\alias{getDifferentiallyExpressedTargetsForRegulons}
\title{Get Differentially Expressed Targets for Regulons}
\usage{
getDifferentiallyExpressedTargetsForRegulons(
  seuratObj,
  regulonNames,
  logFcThreshold,
  grnDf,
  targetCt
)
}
\arguments{
\item{seuratObj}{A Seurat object containing gene expression data.}

\item{regulonNames}{A vector of regulon names to analyze.}

\item{logFcThreshold}{Log fold change threshold for identifying markers.}

\item{grnDf}{Gene regulatory network data frame.}

\item{targetCt}{Target count.}
}
\value{
A list containing markers for each regulon.
}
\description{
This function identifies differentially expressed targets for a given set of regulons
in a Seurat object.
}
