% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{getDifferentiallyExpressedTargetsForRegulonsAllClusters}
\alias{getDifferentiallyExpressedTargetsForRegulonsAllClusters}
\title{Get Differentially Expressed Targets for Regulons in All Clusters}
\usage{
getDifferentiallyExpressedTargetsForRegulonsAllClusters(
  decipher_seurat,
  significant_regulon_deltas_all_clusters,
  regulons_all_clusters,
  flag.normalize.non.log
)
}
\arguments{
\item{decipher_seurat}{A Seurat object containing single-cell RNA-seq data with cluster and condition metadata.}

\item{significant_regulon_deltas_all_clusters}{A list of significant regulon deltas for each cluster, typically obtained from `getSignificantRegulonsAllClusters`.}

\item{regulons_all_clusters}{A list of regulons for each cluster, typically obtained from `getRegulonsAllClusters`.}

\item{flag.normalize.non.log}{A logical flag indicating whether to normalize non-log-transformed data.}
}
\value{
A list where each element corresponds to a cluster and contains the differentially expressed targets for significant regulons in that cluster.
}
\description{
This function identifies differentially expressed target genes for significant regulons in each cluster of a Seurat object.
}
\details{
The function iterates through each unique cluster in the `decipher_seurat` object, subsets the Seurat object for the cluster, normalizes the data if necessary, and identifies differentially expressed targets for significant regulons in that cluster using the `getDifferentiallyExpressedTargetsForRegulons` function.
}
\examples{
\dontrun{
decipher_seurat <- CreateSeuratObject(counts = your_counts_matrix)
significant_regulon_deltas_all_clusters <- getSignificantRegulonsAllClusters(regulon_deltas_all_clusters)
regulons_all_clusters <- getRegulonsAllClusters(filepath, decipher_seurat)
significant_regulon_markers <- getDifferentiallyExpressedTargetsForRegulonsAllClusters(
  decipher_seurat, significant_regulon_deltas_all_clusters, regulons_all_clusters, TRUE)
}

}
