% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_qc.R
\name{getExpressedReceptorsForEachCluster}
\alias{getExpressedReceptorsForEachCluster}
\title{Get Expressed Receptors for Each Cluster}
\usage{
getExpressedReceptorsForEachCluster(seurat_object, L.set)
}
\arguments{
\item{seurat_object}{A Seurat object containing single-cell RNA-seq data.}

\item{L.set}{A set of ligands used for filtering receptors.}
}
\value{
A list where each element corresponds to a cluster and contains the expressed receptors for that cluster.
}
\description{
This function identifies and retrieves expressed receptors for each cluster in a Seurat object.
}
\details{
The function iterates through each unique cluster in the Seurat object and applies the `getFilteredReceptorsForCluster` function to identify receptors expressed in that cluster. The parameter `param_min_receptor_expr_in_cluster` is set to 0.1 to filter receptors based on their expression level.
}
\examples{
\dontrun{
seurat_object <- CreateSeuratObject(counts = your_counts_matrix)
L.set <- c("Ligand1", "Ligand2", "Ligand3")
expressed_receptors <- getExpressedReceptorsForEachCluster(seurat_object, L.set)
}

}
