% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercellular_tailoring.R
\name{getFeatureStatistics}
\alias{getFeatureStatistics}
\title{Calculate Feature Statistics in Seurat Object}
\usage{
getFeatureStatistics(features, seuratObj)
}
\arguments{
\item{features}{A vector of feature (gene) names for which to calculate statistics.}

\item{seuratObj}{A Seurat object containing single-cell RNA sequencing data.}
}
\value{
A data frame with rows representing features and columns representing different statistics:
        cluster, condition, feature, sum of counts, binary counts, number of cells, fraction of cells
        with counts, total counts for the feature in the condition, and fraction of counts for the
        feature in the condition.
}
\description{
This function calculates various statistics for specified features across different clusters
and conditions in a Seurat object. Statistics include the sum of counts, binary counts,
number of cells, fraction of cells with counts, and fractions of total counts for each feature.
}
\examples{
# Assuming 'seurat' is a valid Seurat object and 'features' is a vector of gene names:
stats <- getFeatureStatistics(features, seurat)
}
