% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_qc.R
\name{getFilteredLigands}
\alias{getFilteredLigands}
\title{Filter Ligands Based on Expression Threshold in Clusters}
\usage{
getFilteredLigands(decipher_seurat, L.set, param_min_ligand_expr_in_cluster)
}
\arguments{
\item{decipher_seurat}{A Seurat object containing single-cell RNA-seq data.}

\item{L.set}{A dataframe containing columns `ligand` and `receptor`, which lists
the ligands and their corresponding receptors.}

\item{param_min_ligand_expr_in_cluster}{The minimum fraction of cells (expressed as a decimal)
within any cluster that must express the ligand for it to be considered expressed.}
}
\value{
A vector of ligands that are expressed above the specified threshold across clusters.
}
\description{
This function extracts ligands that are expressed above a specified threshold in one or more
clusters from a Seurat object. It integrates information about ligands and receptors from
a given dataset, computes feature statistics for these molecules, and filters out those ligands
that meet the expression criteria.
}
\examples{
# Assuming 'decipher_seurat' is a Seurat object and 'L.set' is a dataframe
# with ligand and receptor information:
expressed_ligands <- getFilteredLigands(
  decipher_seurat = seurat_object,
  L.set = ligand_receptor_set,
  param_min_ligand_expr_in_cluster = 0.1
)

}
