% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_qc.R
\name{getFilteredReceptorsForCluster}
\alias{getFilteredReceptorsForCluster}
\title{Filter Receptors Based on Expression Threshold within a Specific Cluster}
\usage{
getFilteredReceptorsForCluster(
  decipher_seurat,
  L.set,
  param_min_receptor_expr_in_cluster,
  this_cluster
)
}
\arguments{
\item{decipher_seurat}{A Seurat object containing single-cell RNA-seq data.}

\item{L.set}{A dataframe containing columns `ligand` and `receptor`, which lists
the ligands and their corresponding receptors.}

\item{param_min_receptor_expr_in_cluster}{The minimum fraction of cells (expressed as a decimal)
within the specified cluster that must express the receptor for it to be considered expressed.}

\item{this_cluster}{The cluster within the Seurat object for which receptors are being filtered.}
}
\value{
A vector of receptors that are expressed above the specified threshold within the given cluster.
}
\description{
This function filters out receptors based on their expression levels within a specific cluster
of a Seurat object. It uses a dataset that contains information about both ligands and receptors,
calculates feature statistics for these molecules across the dataset, and returns receptors
that are expressed above a specified threshold in a specified cluster.
}
\examples{
# Assuming 'decipher_seurat' is a Seurat object and 'L.set' is a dataframe
# with ligand and receptor information:
filtered_receptors <- getFilteredReceptorsForCluster(
  decipher_seurat = seurat_object,
  L.set = ligand_receptor_set,
  param_min_receptor_expr_in_cluster = 0.1,
  this_cluster = "Cluster_1"
)

}
