% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercellular_tailoring.R
\name{getForrestLRDatabase}
\alias{getForrestLRDatabase}
\title{Retrieve and Format Ligand-Receptor Data from a Forrest Database}
\usage{
getForrestLRDatabase(forrest.database.filepath)
}
\arguments{
\item{forrest.database.filepath}{A string specifying the path to the CSV file
containing the Forrest ligand-receptor database.}
}
\value{
A data frame with three columns: 'ligand', 'receptor', and 'lr'.
        The 'lr' column is a concatenation of the 'ligand' and 'receptor' columns,
        separated by a hyphen.
}
\description{
This function reads a CSV file containing a ligand-receptor database (typically
from the Forrest database) and reformats it to a specific structure. It selects
ligand and receptor gene symbols from the database and creates a new column combining
these two symbols.
}
\examples{
# Example usage:
filepath <- "path/to/forrest/database.csv"
lr_data <- getForrestLRDatabase(filepath)

}
