% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_data_wrangling.R
\name{getInteractionCorrelationAndSearchSpaceBetweenMethods}
\alias{getInteractionCorrelationAndSearchSpaceBetweenMethods}
\title{Calculate Spearman Correlation and Search Space Between Interaction Methods}
\usage{
getInteractionCorrelationAndSearchSpaceBetweenMethods(method_results_list)
}
\arguments{
\item{method_results_list}{A named list of dataframes, each representing the
results from a different interaction method. Each dataframe should have columns
named `interaction`, `receiver`, and `prioritization_score`, with `prioritization_score.x`
and `prioritization_score.y` being used for merged dataframes to represent scores
from the two methods being compared.}
}
\value{
A list containing two matrices: `spearman` with Spearman correlation coefficients
for all pairs of methods, and `k_matrix` with the corresponding search space size
needed to achieve at least 100 overlapping interactions between each pair of methods.
}
\description{
This function computes the Spearman correlation coefficient for the prioritization
scores between all unique pairs of interaction methods provided. It also determines
the search space (the minimum number of top-ranked interactions) needed to achieve
at least 100 overlaps between the pairs of methods. The results are returned in two
matrices: one for the Spearman correlation coefficients and another for the search
space sizes.
}
\examples{
\dontrun{
  method_list <- list(method1 = df1, method2 = df2)
  correlation_and_search_space <- getInteractionCorrelationAndSearchSpaceBetweenMethods(method_list)
  print(correlation_and_search_space$spearman)
  print(correlation_and_search_space$k_matrix)
}

}
