% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercellular_scoring.R
\name{getInteractionMappingTable}
\alias{getInteractionMappingTable}
\title{Generate Interaction Mapping Table for Ligands and Receptors}
\usage{
getInteractionMappingTable(receptorMatrix, ligandSet)
}
\arguments{
\item{receptorMatrix}{A matrix with receptors as row names.}

\item{ligandSet}{A data frame with columns 'ligand' and 'receptor', representing
ligand-receptor pairs.}
}
\value{
A data frame where each row represents an interaction pair, with columns
for the interaction name (concatenation of ligand and receptor names), ligand name,
and receptor name.
}
\description{
This function creates a mapping table for ligands and receptors based on
a receptor expression matrix and a ligand-receptor set. For each receptor
in the receptor matrix, it identifies the associated ligands from the
ligand-receptor set and constructs a table with the interaction pairs.
}
\examples{
# receptorMatrix is a matrix with receptor genes as row names
# ligandSet is a data frame with ligand-receptor pairs
interactionTable <- getInteractionMappingTable(receptorMatrix, ligandSet)
}
