% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercellular_scoring.R
\name{getInteractionPotentialMatrixForRepresentativeInteractions}
\alias{getInteractionPotentialMatrixForRepresentativeInteractions}
\title{Generate Interaction Potential Matrix for Representative Interactions}
\usage{
getInteractionPotentialMatrixForRepresentativeInteractions(
  data_this_cluster_downsampled_receptors,
  selected_lr_pairs,
  interaction_potentials_matrix_this_cluster,
  cytosig_ligands
)
}
\arguments{
\item{data_this_cluster_downsampled_receptors}{Matrix of receptor expression data
for a downsampled cluster.}

\item{selected_lr_pairs}{Data frame of selected ligand-receptor pairs.}

\item{interaction_potentials_matrix_this_cluster}{Matrix of interaction potentials for the cluster.}

\item{cytosig_ligands}{Vector of prioritized benchmarking ligands used to determine
representative interactions in many-to-one clusters.}
}
\value{
A matrix that combines the interaction potentials for one-to-one interactions
        and representative interactions for many-to-one clusters.
}
\description{
This function analyzes interaction potentials for a specific cluster and categorizes
interactions between receptors and ligands into one-to-one and many-to-one types.
It then identifies representative interactions for the many-to-one category based on
clustering of interaction potentials. This helps in reducing complexity and focusing
on representative interaction dynamics within the cluster.
}
\examples{
# Assuming the required matrices and data frames are predefined:
interaction_matrix <- getInteractionPotentialMatrixForRepresentativeInteractions(
  data_this_cluster_downsampled_receptors = receptor_data,
  selected_lr_pairs = lr_pairs,
  interaction_potentials_matrix_this_cluster = interaction_matrix,
  cytosig_ligands = benchmarking_ligands
)

}
