% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercellular_scoring.R
\name{getInteractionPotentialMatrixForRepresentativeInteractionsAllClusters}
\alias{getInteractionPotentialMatrixForRepresentativeInteractionsAllClusters}
\title{Get Interaction Potential Matrix for Representative Interactions in All Clusters}
\usage{
getInteractionPotentialMatrixForRepresentativeInteractionsAllClusters(
  decipher_seurat,
  L_set_relevant_features_all_clusters,
  filtered_interaction_potentials_matrix_all_clusters,
  cytosig_ligands,
  flag.normalize.non.log
)
}
\arguments{
\item{decipher_seurat}{A Seurat object containing single-cell RNA-seq data with cluster and condition metadata.}

\item{L_set_relevant_features_all_clusters}{A list of relevant ligand-receptor features for each cluster.}

\item{filtered_interaction_potentials_matrix_all_clusters}{A list of filtered interaction potentials matrices for each cluster.}

\item{cytosig_ligands}{A vector of ligands involved in cytosignaling.}

\item{flag.normalize.non.log}{A logical flag indicating whether to normalize non-log-transformed data.}
}
\value{
A list where each element corresponds to a cluster and contains the interaction potential matrix for representative interactions in that cluster.
}
\description{
This function calculates the interaction potential matrix for representative interactions in each cluster of a Seurat object.
}
\details{
The function iterates through each unique cluster in the `decipher_seurat` object, subsets the Seurat object for the cluster, normalizes the data if necessary, and calculates the interaction potential matrix for representative interactions using the `getInteractionPotentialMatrixForRepresentativeInteractions` function.
}
\examples{
\dontrun{
decipher_seurat <- CreateSeuratObject(counts = your_counts_matrix)
L_set_relevant_features_all_clusters <- getRelevantFeaturesForEachCluster(L.set, expressed_ligands, expressed_receptors_all_clusters)
filtered_interaction_potentials_matrix_all_clusters <- filterIntPotByDeltas(interaction_potentials_matrix_all_clusters, interaction_deltas_all_clusters)
cytosig_ligands <- c("Ligand1", "Ligand2")
interaction_potentials_matrix_clusters_all_clusters <- getInteractionPotentialMatrixForRepresentativeInteractionsAllClusters(
  decipher_seurat, L_set_relevant_features_all_clusters, filtered_interaction_potentials_matrix_all_clusters, cytosig_ligands, TRUE)
}

}
