% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercellular_scoring.R
\name{getInteractionPotentialsMatrixAllClusters}
\alias{getInteractionPotentialsMatrixAllClusters}
\title{Get Interaction Potentials Matrix for All Clusters}
\usage{
getInteractionPotentialsMatrixAllClusters(
  decipher_seurat,
  decipher_seurat_this_cluster,
  L_set_relevant_features_all_clusters,
  flag.normalize.non.log
)
}
\arguments{
\item{decipher_seurat}{A Seurat object containing single-cell RNA-seq data with cluster and condition metadata.}

\item{decipher_seurat_this_cluster}{A Seurat object subset for a specific cluster.}

\item{L_set_relevant_features_all_clusters}{A list of relevant ligand-receptor features for each cluster.}

\item{flag.normalize.non.log}{A logical flag indicating whether to normalize non-log-transformed data.}
}
\value{
A list where each element corresponds to a cluster and contains the interaction potentials matrix for that cluster.
}
\description{
This function calculates the interaction potentials matrix for each cluster in a Seurat object.
}
\details{
The function iterates through each unique cluster in the `decipher_seurat` object, subsets the Seurat object for the cluster, normalizes the data if necessary, and calculates the interaction potentials matrix for each cluster using the `getInteractionPotentialsMatrixThisCluster` function.
}
\examples{
\dontrun{
decipher_seurat <- CreateSeuratObject(counts = your_counts_matrix)
L_set_relevant_features_all_clusters <- getRelevantFeaturesForEachCluster(L.set, expressed_ligands, expressed_receptors_all_clusters)
interaction_potentials_matrix <- getInteractionPotentialsMatrixAllClusters(
  decipher_seurat, decipher_seurat_this_cluster, L_set_relevant_features_all_clusters, TRUE)
}

}
