% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercellular_scoring.R
\name{getInteractionPotentialsMatrixThisCluster}
\alias{getInteractionPotentialsMatrixThisCluster}
\title{Compute Interaction Potentials Matrix for a Specific Cluster}
\usage{
getInteractionPotentialsMatrixThisCluster(
  seurat_obj,
  seurat_obj_this_cluster_ds,
  selected_lr_pairs
)
}
\arguments{
\item{seurat_obj}{A Seurat object containing the full dataset.}

\item{seurat_obj_this_cluster_ds}{A downsampled Seurat object for the specific cluster.}

\item{selected_lr_pairs}{A dataframe specifying ligand-receptor pairs to consider,
which must have columns 'ligand' and 'receptor'.}
}
\value{
An interaction potentials matrix where rows represent receptors and
        columns correspond to interaction potentials derived from case-control comparisons
        for the ligands.
}
\description{
This function calculates the interaction potentials between ligands and receptors
within a specific cluster. It separates the case and control conditions in the
main Seurat object, computes the mean expression of selected ligands for both
conditions, and then uses these means along with the expression of receptors
in a downsampled cluster-specific Seurat object to compute an interaction
potentials matrix.
}
\examples{
# Assuming 'full_seurat' is the full Seurat object, 'cluster_seurat_ds' is the downsampled Seurat object for a cluster,
# and 'lr_pairs' is a dataframe of selected ligand-receptor pairs:
interaction_matrix <- getInteractionPotentialsMatrixThisCluster(
  seurat_obj = full_seurat,
  seurat_obj_this_cluster_ds = cluster_seurat_ds,
  selected_lr_pairs = lr_pairs
)

}
