% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_data_wrangling.R
\name{getLigandReceptorDiffExprMarkersByCt}
\alias{getLigandReceptorDiffExprMarkersByCt}
\title{Get Differential Expression Markers for Ligands and Receptors by Cell Type}
\usage{
getLigandReceptorDiffExprMarkersByCt(lr.marker.list)
}
\arguments{
\item{lr.marker.list}{A list where each element represents a different cell type
and contains a data frame of ligand-receptor differential expression markers for that cell type.
Each data frame should have row names set to gene names and contain columns for
average log2 fold change (avg_log2FC), and percentage of cells expressing the gene in
two different conditions (pct.1, pct.2).}
}
\value{
A data frame consolidating all input differential expression markers with columns:
\itemize{
\item \code{avg_log2FC}: Average log2 fold change of expression.
\item \code{cluster}: Cell type or cluster.
\item \code{gene}: Gene name.
\item \code{pct.1}: Percentage of cells in condition 1 expressing the gene.
\item \code{pct.2}: Percentage of cells in condition 2 expressing the gene.
}
}
\description{
This function aggregates differential expression markers for ligands and receptors
across different cell types specified in the input list. It constructs a data frame
that includes average log fold change, the cluster or cell type, the gene name,
and the percentage of cells in two conditions expressing the gene.
}
\examples{
# Assuming 'lr_marker_list' is pre-defined with the appropriate structure
results <- getLigandReceptorDiffExprMarkersByCt(lr_marker_list)

}
