% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercellular_scoring.R
\name{getManyToOneInteractions}
\alias{getManyToOneInteractions}
\title{Extract Many-to-One Interactions from a Mapping Table}
\usage{
getManyToOneInteractions(mappingTable)
}
\arguments{
\item{mappingTable}{A data frame representing a mapping table with at least
two columns: 'receptor' and 'interaction'.}
}
\value{
A vector of interaction names that represent many-to-one ligand-receptor
pairs from the mapping table.
}
\description{
This function identifies many-to-one interactions in a mapping table, where each
receptor is associated with multiple ligands. It returns the interactions
that have a receptor linked to more than one ligand.
}
\examples{
# mappingTable is a data frame with columns 'interaction' and 'receptor'
manyToOneInteractions <- getManyToOneInteractions(mappingTable)
}
