% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_data_wrangling.R
\name{getNumberOfOutliers}
\alias{getNumberOfOutliers}
\title{Count the Number of Outliers in a Dataframe}
\usage{
getNumberOfOutliers(df)
}
\arguments{
\item{df}{A dataframe that must contain a column named 'prioritization_score'.}
}
\value{
A vector with three elements, each representing the count of values
that are more than one, two, and three standard deviations above the mean, respectively.
}
\description{
This function calculates the number of outliers in the 'prioritization_score'
column of a dataframe. An outlier is defined as a value that is more than a
certain number of standard deviations away from the mean. This function counts
how many values exceed one, two, and three standard deviations above the mean.
}
\examples{
\dontrun{
  # Assuming 'your_dataframe' has a column 'prioritization_score'
  outlier_counts <- getNumberOfOutliers(your_dataframe)
  print(outlier_counts)
}

}
