% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_data_wrangling.R
\name{getPredictionsResponsesForMethods}
\alias{getPredictionsResponsesForMethods}
\title{Extract Predictions and Responses for Comparison Across Methods}
\usage{
getPredictionsResponsesForMethods(
  results_to_compare,
  cytosig_significance,
  L.set,
  seurat_object_oi,
  output_figures_filepath
)
}
\arguments{
\item{results_to_compare}{A named list of dataframes, each containing the prediction
results of a different method. Each dataframe should include columns for ligand
names and scores.}

\item{cytosig_significance}{A dataframe containing significance data for genes,
used as the basis for response data in the comparison.}

\item{L.set}{A dataframe specifying the set of ligands considered in the analysis,
used to filter the predictions and responses.}

\item{seurat_object_oi}{A Seurat object or equivalent containing gene expression
data, used to further filter the genes considered in the analysis based on their
presence in the dataset.}
}
\value{
A list containing two elements: `predictions` and `responses`, each a list
where each element corresponds to the prediction vectors or response vectors for
a method, respectively.
}
\description{
This function compares predictions from various methods against response data
available in a significance dataset. It processes each method's predictions, aligns
them with response data based on matching genes, and then generates plots to
visualize the comparison. The function aims to assess how closely the predictions
of each method match actual response data, considering only significant genes and
those within a specified ligand set.
}
\examples{
\dontrun{
  # Example usage
  result <- getPredictionsResponsesForMethods(results_to_compare,
                                              cytosig_significance,
                                              L.set,
                                              seurat_object_oi)
}

}
