% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decipher_scoring.R
\name{getRandomForestWeightsAllClustersWParamPairings}
\alias{getRandomForestWeightsAllClustersWParamPairings}
\title{Get Random Forest Weights for All Clusters}
\usage{
getRandomForestWeightsAllClustersWParamPairings(
  decipher_seurat,
  significant_regulon_deltas_all_clusters,
  regulon_scores_all_clusters,
  interaction_potentials_matrix_clusters_all_clusters,
  L_set_relevant_features_all_clusters,
  flag.normalize.non.log,
  paramPairings
)
}
\arguments{
\item{decipher_seurat}{A Seurat object containing single-cell RNA-seq data with cluster and condition metadata.}

\item{significant_regulon_deltas_all_clusters}{A list of significant regulon deltas for each cluster.}

\item{regulon_scores_all_clusters}{A list of regulon scores for each cluster.}

\item{interaction_potentials_matrix_clusters_all_clusters}{A list of interaction potentials matrices for each cluster.}

\item{L_set_relevant_features_all_clusters}{A list of relevant ligand-receptor features for each cluster.}

\item{flag.normalize.non.log}{A logical flag indicating whether to normalize non-log-transformed data.}

\item{paramPairings}{}
}
\value{
A list where each element corresponds to a cluster and contains the random forest importance results for that cluster.
}
\description{
This function calculates random forest weights for all clusters in a Seurat object using significant regulon deltas, regulon scores, interaction potentials, and relevant ligand-receptor features.
}
\details{
The function iterates through each unique cluster in the Seurat object, subsets the Seurat object for the cluster, normalizes the data if necessary, and calculates random forest importance scores for the transcription factors using the `randomForest` package. A progress bar is displayed to show the progress of the computation.
}
\examples{
\dontrun{
decipher_seurat <- CreateSeuratObject(counts = your_counts_matrix)
significant_regulon_deltas_all_clusters <- getSignificantRegulonsAllClusters(regulon_deltas_all_clusters)
regulon_scores_all_clusters <- getRegulonScoresAllClusters(capped_regulons_all_clusters, decipher_seurat)
interaction_potentials_matrix_clusters_all_clusters <- getInteractionPotentialMatrixForRepresentativeInteractionsAllClusters(
  decipher_seurat, L_set_relevant_features_all_clusters, filtered_interaction_potentials_matrix_all_clusters, cytosig_ligands, TRUE)
rf_results_all_clusters <- getRandomForestWeightsAllClusters(
  decipher_seurat, significant_regulon_deltas_all_clusters, regulon_scores_all_clusters, interaction_potentials_matrix_clusters_all_clusters, L_set_relevant_features_all_clusters, TRUE)
}

}
