% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intracellular_tailoring.R
\name{getRegulon}
\alias{getRegulon}
\title{Retrieve and Optionally Trim Gene Regulatory Network (GRN) for a Specific Cluster}
\usage{
getRegulon(output_filepath, this_cluster, flag.co.grn = TRUE)
}
\arguments{
\item{output_filepath}{The base directory where the CellOracle data and GRN files are stored.}

\item{this_cluster}{The name of the cluster for which the GRN is to be retrieved.}

\item{flag.co.grn}{Logical flag indicating whether to trim the GRN using CellOracle criteria.
If TRUE, the GRN is trimmed according to specified parameters.}
}
\value{
A data frame representing the GRN for the specified cluster. If no GRN file is found,
        a warning is issued and the function returns `NULL`.
}
\description{
This function loads a gene regulatory network (GRN) from a CSV file specific to a given cluster.
It optionally trims the GRN based on thresholds specified for p-values, the number of top edges,
and the minimum number of targets, using the CellOracle toolkit if enabled.
}
\examples{
# Assuming the necessary files are in 'path/to/output', and you're interested in 'Cluster_1':
regulon_data <- getRegulon(
  output_filepath = "path/to/output",
  this_cluster = "Cluster_1",
  flag.co.grn = TRUE
)

}
