% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intracellular_scoring.R
\name{getRegulonDeltas}
\alias{getRegulonDeltas}
\title{Calculate Differential Regulation for Regulons}
\usage{
getRegulonDeltas(regulonMatrix, classVector)
}
\arguments{
\item{regulonMatrix}{A numeric matrix where rows represent regulons (or pathways)
and columns represent different samples.}

\item{classVector}{A character vector indicating the class for each sample.
It should have the same length as the number of columns in `regulonMatrix`.
The function calculates the difference between the medians of the 'case' class
and other classes.}
}
\value{
A data frame where each row corresponds to a regulon and contains
the differential regulation value (`deltaPagoda`) and the name of the regulon.
}
\description{
This function computes the difference in median values for each regulon
(row) in a matrix between two classes. The classes are defined in the
provided class vector, distinguishing between 'case' and other categories.
}
\examples{
# regulonMatrix is a matrix of regulon activities
# classVector is a vector of class labels for each sample
deltas <- getRegulonDeltas(regulonMatrix, classVector)
}
