% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intracellular_scoring.R
\name{getRegulonDeltasAllClustersWParamPairings}
\alias{getRegulonDeltasAllClustersWParamPairings}
\title{Get Regulon Deltas for All Clusters}
\usage{
getRegulonDeltasAllClustersWParamPairings(
  regulon_scores_all_clusters,
  decipher_seurat,
  paramPairings
)
}
\arguments{
\item{regulon_scores_all_clusters}{A list of regulon scores for each cluster, typically obtained from `getRegulonScoresAllClusters`.}

\item{decipher_seurat}{A Seurat object containing single-cell RNA-seq data with cluster and condition metadata.}

\item{paramPairings}{}
}
\value{
A list where each element corresponds to a cluster and contains the regulon deltas for that cluster.
}
\description{
This function calculates the differences (deltas) in regulon scores between conditions for each cluster in a Seurat object.
}
\details{
The function iterates through each unique cluster in the `decipher_seurat` object, subsets the Seurat object for the cluster, normalizes the data if necessary, and calculates the deltas in regulon scores for the conditions in that cluster.
}
\examples{
\dontrun{
regulon_scores_all_clusters <- getRegulonScoresAllClusters(capped_regulons_all_clusters, decipher_seurat)
decipher_seurat <- CreateSeuratObject(counts = your_counts_matrix)
regulon_deltas <- getRegulonDeltasAllClusters(regulon_scores_all_clusters, decipher_seurat)
}

}
