% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intracellular_tailoring.R
\name{getRegulonsAllClusters}
\alias{getRegulonsAllClusters}
\title{Get Regulons for All Clusters}
\usage{
getRegulonsAllClusters(filepath, seurat_object)
}
\arguments{
\item{filepath}{A string representing the path to the file containing regulon data.}

\item{seurat_object}{A Seurat object containing single-cell RNA-seq data.}
}
\value{
A list where each element corresponds to a cluster and contains the regulons for that cluster.
}
\description{
This function retrieves regulons for each cluster in a Seurat object from a specified file.
}
\details{
The function iterates through each unique cluster in the Seurat object and applies the `getRegulon` function to retrieve regulons for that cluster from the specified file. The results are stored in a list, with each element corresponding to a cluster.
}
\examples{
\dontrun{
filepath <- "path/to/regulon/file"
seurat_object <- CreateSeuratObject(counts = your_counts_matrix)
regulons <- getRegulonsAllClusters(filepath, seurat_object)
}

}
