% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intracellular_scoring.R
\name{getRegulonScores}
\alias{getRegulonScores}
\title{Calculate Regulon Scores Using Pagoda2}
\usage{
getRegulonScores(seuratObject, grn_df)
}
\arguments{
\item{seuratObject}{A Seurat object containing single-cell RNA sequencing data.}

\item{grn_df}{A data frame representing the gene regulatory network with columns specifying gene interactions.}
}
\value{
A matrix (RMatrixSC) containing the regulon scores calculated from the Pagoda2 analysis.
}
\description{
This function calculates regulon scores for a given Seurat object and gene regulatory network (GRN) data frame.
It involves creating a Pagoda2 object, preparing the GRN environment, running the Pagoda2 analysis, and processing the results.
}
\note{
The function assumes that the 'RNA' assay and the 'data' slot in the Seurat object are appropriate for the analysis.
It also utilizes several helper functions: 'createAndPreparePagoda2', 'prepGRN', 'prepareGRNEnvironment', and 'processPagodaResults'.
These functions need to be defined and available in the environment.
}
\examples{
# seuratObject is a pre-existing Seurat object
# grn_df is a data frame representing the gene regulatory network
regulonScores <- getRegulonScores(seuratObject, grn_df)

}
