% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intracellular_scoring.R
\name{getRegulonScoresAllClustersWParamPairings}
\alias{getRegulonScoresAllClustersWParamPairings}
\title{Get Regulon Scores for All Clusters}
\usage{
getRegulonScoresAllClustersWParamPairings(
  capped_regulons_all_clusters,
  decipher_seurat,
  paramPairings
)
}
\arguments{
\item{capped_regulons_all_clusters}{A list of capped regulons for each cluster, typically obtained from `capRegulonsAllClusters`.}

\item{decipher_seurat}{A Seurat object containing single-cell RNA-seq data with cluster and condition metadata.}

\item{paramPairings}{}
}
\value{
A list where each element corresponds to a cluster and contains the regulon scores for that cluster.
}
\description{
This function calculates regulon scores for each cluster in a Seurat object based on capped regulons.
}
\details{
The function iterates through each unique cluster in the `decipher_seurat` object, subsets the Seurat object for the cluster, normalizes the data if necessary, and calculates the regulon scores for the capped regulons of that cluster.
}
\examples{
\dontrun{
capped_regulons_all_clusters <- capRegulonsAllClusters(regulons_all_clusters, decipher_seurat, TRUE)
decipher_seurat <- CreateSeuratObject(counts = your_counts_matrix)
regulon_scores <- getRegulonScoresAllClusters(capped_regulons_all_clusters, decipher_seurat)
}

}
