% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_qc.R
\name{getRelevantFeaturesForEachCluster}
\alias{getRelevantFeaturesForEachCluster}
\title{Get Relevant Features for Each Cluster}
\usage{
getRelevantFeaturesForEachCluster(
  L.set,
  expressed_ligands,
  expressed_receptors_all_clusters
)
}
\arguments{
\item{L.set}{A data frame containing ligand-receptor pairs.}

\item{expressed_ligands}{A vector of expressed ligands.}

\item{expressed_receptors_all_clusters}{A list of expressed receptors for each cluster, typically obtained from `getExpressedReceptorsForEachCluster`.}
}
\value{
A list where each element corresponds to a cluster and contains the relevant ligand-receptor features for that cluster.
}
\description{
This function identifies and retrieves relevant ligand-receptor features for each cluster based on expressed ligands and receptors.
}
\details{
The function iterates through each cluster in the `expressed_receptors_all_clusters` list, filters the `L.set` data frame to include only those ligand-receptor pairs where the receptor is expressed in the cluster and the ligand is in the `expressed_ligands` vector. The filtered results are then stored in a list, with each element corresponding to a cluster.
}
\examples{
\dontrun{
L.set <- data.frame(ligand = c("Ligand1", "Ligand2"), receptor = c("Receptor1", "Receptor2"))
expressed_ligands <- c("Ligand1", "Ligand3")
expressed_receptors_all_clusters <- list(Cluster1 = c("Receptor1"), Cluster2 = c("Receptor2"))
relevant_features <- getRelevantFeaturesForEachCluster(L.set, expressed_ligands, expressed_receptors_all_clusters)
}

}
