% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intracellular_scoring.R
\name{getSignificantRegulons}
\alias{getSignificantRegulons}
\title{Identify Significant Regulons Based on Delta Values}
\usage{
getSignificantRegulons(regulon_deltas_this_cluster)
}
\arguments{
\item{regulon_deltas_this_cluster}{A data frame containing delta values for regulons,
which should include columns 'class' (indicating whether the regulon is 'real' or 'random')
and 'deltaPagoda' (the delta values to be analyzed).}
}
\value{
A data frame of regulons classified as 'real' that have deltaPagoda values falling outside
        the upper and lower thresholds defined by the random regulons' density.
}
\description{
This function filters regulons by identifying significant changes in their delta values compared to a random distribution.
It computes the density of delta values for random regulons, then uses the 2.5th and 97.5th percentiles as thresholds
to identify real regulons with delta values that are significantly higher or lower than these thresholds.
}
\examples{
# Assuming 'regulon_deltas_cluster' is a data frame with the necessary structure:
significant_regulons <- getSignificantRegulons(regulon_deltas_cluster)

}
