% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intracellular_scoring.R
\name{getSignificantRegulonsAllClusters}
\alias{getSignificantRegulonsAllClusters}
\title{Get Significant Regulons for All Clusters}
\usage{
getSignificantRegulonsAllClusters(regulon_deltas_all_clusters)
}
\arguments{
\item{regulon_deltas_all_clusters}{A list of regulon deltas for each cluster, typically obtained from `getRegulonDeltasAllClusters`.}
}
\value{
A list where each element corresponds to a cluster and contains the significant regulon deltas for that cluster.
}
\description{
This function identifies significant regulon deltas for each cluster.
}
\details{
The function iterates through each cluster in the `regulon_deltas_all_clusters` list, applies the `getSignificantRegulons` function to identify significant regulon deltas, and stores the results in a list with each element corresponding to a cluster.
}
\examples{
\dontrun{
regulon_deltas_all_clusters <- getRegulonDeltasAllClusters(regulon_scores_all_clusters, decipher_seurat)
significant_regulons <- getSignificantRegulonsAllClusters(regulon_deltas_all_clusters)
}

}
