% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_data_wrangling.R
\name{getTopRegulonsAndCts}
\alias{getTopRegulonsAndCts}
\title{Extract Top Regulons and Corresponding Cell Types}
\usage{
getTopRegulonsAndCts(regulon_deltas_by_cluster)
}
\arguments{
\item{regulon_deltas_by_cluster}{A list of dataframes, each representing regulon
delta values for a specific cell type or cluster. Each dataframe is expected to have
been structured to be compatible with `filter_real` and `get_top_regulons` functions.}
}
\value{
A list with two elements:
\itemize{
  \item \code{top_regulons}: A vector of unique top regulon names across all clusters,
  filtered based on a delta threshold.
  \item \code{cts}: The names of clusters (cell types) corresponding to the identified
  top regulons.
}
}
\description{
This function processes a list of dataframes containing regulon delta values by cluster,
filters for "real" entries, identifies the top regulons based on a specified delta threshold,
and extracts the corresponding cell types. It returns a list containing the unique top regulons
across all clusters and the names of clusters (cell types) associated with these top regulons.
}
\examples{
\dontrun{
  regulon_deltas <- list(
    Cluster1 = data.frame(deltaPagoda = c(4, 2, 1), name = c("RegulonA", "RegulonB", "RegulonC")),
    Cluster2 = data.frame(deltaPagoda = c(5, 1, 3), name = c("RegulonA", "RegulonD", "RegulonE"))
  )
  result <- getTopRegulonsAndCts(regulon_deltas)
  print(result$top_regulons)
  print(result$cts)
}

}
\seealso{
\code{\link{filter_real}}, \code{\link{get_top_regulons}}
}
