% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_data_wrangling.R
\name{getTopRowsByMethod}
\alias{getTopRowsByMethod}
\title{Select Top Rows Based on Method and Score}
\usage{
getTopRowsByMethod(
  method_results,
  n_rows = 1,
  negative = FALSE,
  selected_method = NULL
)
}
\arguments{
\item{method_results}{Data frame containing method results, including columns for 'receiver', 'method',
and 'prioritization_score'.}

\item{n_rows}{The number of top rows to select based on the score (default is 1).}

\item{negative}{Logical flag indicating whether to select for negative scores (default is FALSE, select
for positive scores).}

\item{selected_method}{Optionally specify a method to filter by in negative selection mode.}
}
\value{
A data frame of the selected top rows with an additional column 'label' that indicates
        whether the selection was for the top positive or top negative scores.
}
\description{
This function selects the top rows from a dataset based on the specified method and scoring criteria.
It can operate in either a positive or negative mode, determined by the `negative` flag.
In positive mode, it selects the top positive scores. In negative mode, and if a specific method is
selected, it picks the top negative scores for that method.
}
\examples{
# Example DataFrame
df <- data.frame(
  receiver = c("A", "B", "A", "B"),
  method = c("Method1", "Method1", "Method2", "Method2"),
  prioritization_score = c(0.9, 0.85, 0.95, 0.80)
)
# Get top positive rows by default
top_positive <- getTopRowsByMethod(df)
# Get top negative rows for a specific method
top_negative <- getTopRowsByMethod(df, negative = TRUE, selected_method = "Method2")

}
