% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{listOfDFsRenameColumn}
\alias{listOfDFsRenameColumn}
\title{Rename Column in a List of Data Frames}
\usage{
listOfDFsRenameColumn(decipher_scores_by_regulon, original_name, new_name)
}
\arguments{
\item{decipher_scores_by_regulon}{A list where each element is a data frame.}

\item{original_name}{The original name of the column to be renamed.}

\item{new_name}{The new name for the column.}
}
\value{
A list of data frames with the specified column renamed in each data frame.
}
\description{
This function renames a specified column in each data frame within a list of data frames.
}
\examples{
# Assuming 'listDFs' is a list of data frames and you want to rename column 'oldCol' to 'newCol':
renamedList <- listOfDFsRenameColumn(listDFs, "oldCol", "newCol")
}
