% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{loadCytosigLigands}
\alias{loadCytosigLigands}
\title{Load Cytosig Ligands and Convert for Species if Necessary}
\usage{
loadCytosigLigands(reference_filepath, species)
}
\arguments{
\item{reference_filepath}{The directory path where the Cytosig ligands RDS file is stored.}

\item{species}{A string specifying the species, which affects whether a conversion is performed.
The function currently converts the data if "mouse" is specified.}
}
\value{
A data frame or list of Cytosig ligands, potentially converted to mouse gene symbols if
        specified by the species parameter.
}
\description{
This function loads a set of cytotoxic signaling (Cytosig) ligands from a pre-saved RDS file.
Initially, it assumes the ligands are formatted for human. If the species specified is "mouse",
it converts the ligand symbols from human to mouse using a conversion function. This function is
designed to be adaptable for use with mouse data by converting human gene symbols to mouse.
}
\examples{
# Load Cytosig ligands for human (no conversion):
ligands_human <- loadCytosigLigands("/path/to/data", "human")

# Load and convert Cytosig ligands for mouse:
ligands_mouse <- loadCytosigLigands("/path/to/data", "mouse")

}
