% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{loadEnrichrDatabase}
\alias{loadEnrichrDatabase}
\title{Load Enrichr Database Based on Species}
\usage{
loadEnrichrDatabase(reference_filepath, species)
}
\arguments{
\item{reference_filepath}{The directory path where the Enrichr database RDS files are stored.}

\item{species}{A string specifying the species, which determines the database file to load.
Valid options are "human" and "mouse".}
}
\value{
The loaded Enrichr database as an R object, from the RDS file specific to the given species.
}
\description{
This function loads an Enrichr database from a specified path depending on the species indicated.
Currently, it supports loading pre-saved RDS files for human and mouse species. The function
handles species-specific database files, loading the appropriate database based on the species
parameter.
}
\examples{
# Load the Enrichr database for human:
enrichr_db_human <- loadEnrichrDatabase("/path/to/databases", "human")

# Load the Enrichr database for mouse:
enrichr_db_mouse <- loadEnrichrDatabase("/path/to/databases", "mouse")

}
