% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{loadLSet}
\alias{loadLSet}
\title{Load Ligand-Receptor Set from Reference Database}
\usage{
loadLSet(reference_filepath, species)
}
\arguments{
\item{reference_filepath}{The file path to the directory containing the ligand-receptor CSV file.}

\item{species}{A string indicating the species of the dataset, currently supports "mouse" for
species-specific conversion.}
}
\value{
A dataframe of the ligand-receptor set, potentially converted for mouse, including a new
        column 'interaction' that concatenates ligand and receptor names.
}
\description{
This function loads a ligand-receptor interaction set from a specified CSV file within a given directory.
It formats the dataset to include unique ligand-receptor pairs and, if specified, converts the set
for use with mouse data. The function supports customization for different species, currently handling
conversion specific to mouse.
}
\examples{
# Load a ligand-receptor set for mouse from a specified directory:
L.set <- loadLSet("/path/to/directory", "mouse")

# Load a ligand-receptor set without species conversion:
L.set <- loadLSet("/path/to/directory", "human")

}
