% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mapConditionsInSeurat}
\alias{mapConditionsInSeurat}
\title{Map Conditions in Seurat Object}
\usage{
mapConditionsInSeurat(seuratObj, metadataName, caseCondition, controlCondition)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{metadataName}{The name of the metadata column in the Seurat object to map from.}

\item{caseCondition}{The condition value to be mapped to "case".}

\item{controlCondition}{The condition value to be mapped to "control".}
}
\value{
The Seurat object with the mapped conditions added to its metadata.
}
\description{
This function maps the conditions in the metadata of a Seurat object according to a provided mapping.
}
