% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing_meta_cells.R
\name{metaCellModule}
\alias{metaCellModule}
\title{Meta Cell Module Processing}
\usage{
metaCellModule(seurat_object, min_meta_cells)
}
\arguments{
\item{seurat_object}{A Seurat object containing single-cell RNA-seq data.}

\item{min_meta_cells}{The minimum number of meta cells required for processing.}
}
\value{
A Seurat object that has been transformed into a pseudo-bulk format and normalized.
}
\description{
This function orchestrates several steps to process a Seurat object through meta cell analysis.
It calculates the suggested number of meta cell neighbours based on a minimum meta cell threshold,
generates meta cell matrices, creates a pseudo-bulk Seurat object from these matrices, and
normalizes the data. The normalization is done using a 'RC' method with a specific scale factor.
}
\examples{
# Assuming 'seurat_object' is a Seurat object prepared for meta cell analysis
result_seurat <- metaCellModule(seurat_object, 50)

}
