% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_visualizations.R
\name{pairwise_comparison_plot}
\alias{pairwise_comparison_plot}
\title{Pairwise Comparison of Prioritisation Scores}
\usage{
pairwise_comparison_plot(
  df1,
  df2,
  method1,
  method2,
  overlap_max = NA,
  use_rank = FALSE,
  re_rank = FALSE
)
}
\arguments{
\item{df1}{A data frame containing prioritisation scores from the first method.}

\item{df2}{A data frame containing prioritisation scores from the second method.}

\item{method1}{A character string representing the name of the first method.}

\item{method2}{A character string representing the name of the second method.}

\item{overlap_max}{An integer representing a cap on the number of overlapping interactions such that the top overlap_max interactions from each method are selected}
}
\value{
A png saved in the figures folder displaying the correlation of prioritisation scores.
}
\description{
This function takes in two data frames, conducts an inner join based on
interaction and receiver columns, calculates the Spearman correlation, and
produces a ggplot object displaying the correlation.
}
\examples{
# pairwise_comparison_plot(nichenet_for_correlation, decipher_pre_processed, "NicheNet", "Decipher")
}
