% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_visualizations.R
\name{plotBubble}
\alias{plotBubble}
\title{Create a bubble plot with gradient fill and customizable aesthetics}
\usage{
plotBubble(
  df,
  color.var,
  size.var,
  stroke.var,
  plot.position,
  col.min.val,
  col.max.val,
  plot.title,
  x_lab,
  y_lab,
  x_var
)
}
\arguments{
\item{df}{Data frame containing the variables for plotting.}

\item{color.var}{A string specifying the column in `df` used for color gradient.}

\item{size.var}{A string specifying the column in `df` used for point size.}

\item{stroke.var}{A string specifying the stroke width of points.}

\item{plot.position}{A string indicating the plot position ('middle', 'right', or other).}

\item{col.min.val}{The minimum value for the color gradient scale.}

\item{col.max.val}{The maximum value for the color gradient scale.}

\item{plot.title}{String for the plot's title.}

\item{x_lab}{Label for the x-axis.}

\item{y_lab}{Label for the y-axis.}

\item{x_var}{A string specifying the x-axis variable.}
}
\value{
A ggplot object representing the bubble plot.
}
\description{
This function generates a bubble plot using ggplot2, with various customizable aesthetic features.
It plots an interaction variable against a specified x-axis variable, coloring the points based on a color variable and scaling by size.
Additional customizations include gradient color filling, position of the plot, and optional axis adjustments.
}
\examples{
df <- data.frame(interaction = rnorm(10), A = rnorm(10), B = sample(1:100, 10), C = sample(1:5, 10, replace = TRUE))
plotBubble(df, color.var="A", size.var="B", stroke.var="C", plot.position="middle",
           col.min.val=-2, col.max.val=2, plot.title="My Bubble Plot", x_lab="X Axis", y_lab="Y Axis", x_var="A")
}
