% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_visualizations.R
\name{plotDecipherPrioritizedMap}
\alias{plotDecipherPrioritizedMap}
\title{Plot Decipher Prioritized Map}
\usage{
plotDecipherPrioritizedMap(
  dataset_path,
  top_n,
  selected_receiver_cells = NULL,
  sc_feature_statistics = FALSE,
  primary_ct = NULL,
  split_by_direction = FALSE,
  direction = c("pos", "neg"),
  dataset_name,
  abs_decipher_plot_limit = NULL,
  priority_receiver_cells = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{dataset_path}{Character. The path to the dataset directory containing the data files.}
}
\value{
This function does not return a value. It saves a plot and a CSV file in the specified dataset path.
}
\description{
This function reads and processes data to create and save a prioritized interaction map
for ligand-receptor interactions using the Decipher package.
}
\examples{
\dontrun{
plotDecipherPrioritizedMap("path/to/dataset")
}

}
