% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_visualizations.R
\name{plotDecipherPrioritizedMap_v2}
\alias{plotDecipherPrioritizedMap_v2}
\title{Plot Decipher Prioritized Map}
\usage{
plotDecipherPrioritizedMap_v2(
  dataset_path,
  receiver_cell_type = NULL,
  output_filename,
  log_transform,
  slice_n = 5,
  return_plot_object = FALSE
)
}
\arguments{
\item{dataset_path}{Character. The path to the dataset directory containing the data files.}

\item{receiver_cell_type}{Character. The receiver cell type to filter interactions by.}

\item{output_filename}{Character. The filename for the saved plot and CSV file.}

\item{log_transform}{log transform}

\item{slice_n}{number of top interactions}

\item{return_plot_object}{}
}
\value{
This function does not return a value. It saves a plot and a CSV file in the specified dataset path.
}
\description{
This function reads and processes data to create and save a prioritized interaction map
for ligand-receptor interactions using the Decipher package.
}
\examples{
\dontrun{
plotDecipherPrioritizedMap("path/to/dataset", "CD4_T", "decipher_plot_prioritized")
}

}
