% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_visualizations.R
\name{plotInteractionCorrelation}
\alias{plotInteractionCorrelation}
\title{Plot Spearman Correlation Matrix}
\usage{
plotInteractionCorrelation(spearman_matrix, dataset_name, file_path)
}
\arguments{
\item{spearman_matrix}{A square matrix containing Spearman correlation coefficients
between different methods. It is assumed that the matrix is symmetric.}

\item{dataset_name}{A character string representing the name of the dataset,
which will be used as part of the filename for the saved heatmap image.}

\item{file_path}{The directory path where the heatmap image file will be saved.}
}
\value{
Invisibly returns NULL, as the main purpose is side-effect (image file creation).
}
\description{
Generates a heatmap from a Spearman correlation matrix for interaction data between methods.
The function modifies the input matrix to zero out the lower triangular part for better
visualization of the symmetric matrix and saves the heatmap as a PNG file.
}
\details{
The function zeros out the lower triangular part of the spearman_matrix
to focus on the upper triangular part in the heatmap. It then saves the heatmap
as a PNG file named using the `dataset_name` in the specified `file_path`.
The heatmap visualizes the absolute values of the Spearman correlation coefficients
with color gradients.
}
\examples{
\dontrun{
  # Assuming `spearman_matrix` is your Spearman correlation matrix
  plotInteractionCorrelation(spearman_matrix, "MyDataset", "/path/to/save")
}

}
