% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_visualizations.R
\name{plotSearchSpace}
\alias{plotSearchSpace}
\title{Plot Search Space Matrix}
\usage{
plotSearchSpace(k_matrix, dataset_name, file_path)
}
\arguments{
\item{k_matrix}{A symmetric matrix where each element represents the minimum number
of top interactions required to achieve at least 100 overlapping interactions
between the methods corresponding to the row and column.}

\item{dataset_name}{A character string representing the name of the dataset, used
to generate the filename of the output PNG file.}

\item{file_path}{The directory path where the output PNG file will be saved.}
}
\value{
A heatmap visualization of the k matrix is saved as a PNG file at the
specified location. The function itself invisibly returns NULL.
}
\description{
Creates a heatmap visualization of the search space (k) matrix, which indicates
the minimum number of top-ranked interactions needed to achieve a specified number
of overlaps between pairs of methods. The lower triangle of the matrix is masked
to NA to avoid redundancy in the symmetric matrix. The heatmap is saved as a PNG
file.
}
\examples{
\dontrun{
  plotSearchSpace(k_matrix, "my_dataset", "/path/to/save/directory")
}

}
