% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_visualizations.R
\name{plotSpearmanHeatmap}
\alias{plotSpearmanHeatmap}
\title{Plot Spearman Correlation Heatmap}
\usage{
plotSpearmanHeatmap(
  spearman_matrix,
  labels = NULL,
  output_filepath,
  manuscript = FALSE
)
}
\arguments{
\item{spearman_matrix}{A matrix of Spearman correlation coefficients.}

\item{labels}{(Optional) A matrix of labels corresponding to the rows and columns
of the `spearman_matrix`. If provided, these labels are used to annotate the heatmap
cells. Note: This parameter is implicitly inferred from the context but not explicitly
defined in the function signature.}

\item{dataset}{A character string indicating the dataset name, used in generating
the output file name.}
}
\description{
Creates a heatmap visualization of the Spearman correlation matrix. If labels are
provided, they are used to annotate the heatmap; otherwise, a simpler heatmap is
generated. The heatmap is saved as a PNG file in a predetermined directory, with
the filename reflecting whether labels were used.
}
\details{
The function generates two types of heatmaps based on the presence of
`labels`. If `labels` is `NULL`, a heatmap without cell annotations is created. If
`labels` are provided, they are assumed to come from a `label_matrix` (which must be
available in the scope where the function is called) and are used to annotate the
heatmap cells.

The output PNG file is saved in the "figures" directory, with the filename incorporating
the dataset name and indicating whether the heatmap includes top overlapping
transcription factors (TFs).
}
\examples{
\dontrun{
  # Assuming 'spearman_mat' is your Spearman correlation matrix
  plotSpearmanHeatmap(spearman_mat, "my_dataset")
  # To plot with labels, ensure 'label_matrix' is defined in your environment
  plotSpearmanHeatmap(spearman_mat, "my_dataset", labels=TRUE)
}

}
