% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_data_wrangling.R
\name{prepareLianaForCytosigComparison}
\alias{prepareLianaForCytosigComparison}
\title{Prepare LIANA Data for Cytosig Comparison}
\usage{
prepareLianaForCytosigComparison(liana_df)
}
\arguments{
\item{liana_df}{A dataframe from LIANA analysis containing columns for ligand,
receptor, source cluster, target cluster, and interaction statistic.}
}
\value{
A list of data frames, each corresponding to a different receiver cluster
from the original `liana_df`. Columns are standardized for direct comparison with
Cytosig results, and an additional 'interaction' column is included, combining
ligand and receptor names.
}
\description{
Transforms a LIANA data frame for comparison with Cytosig results. This involves
selecting relevant columns, renaming them for consistency with Cytosig data
standards, and creating an interaction column by concatenating ligand and receptor
names. The data is then split into a list of data frames based on the receiver
cluster, facilitating further comparison or analysis.
}
\examples{
\dontrun{
  # Assuming 'liana_df' is your dataframe obtained from LIANA analysis
  liana_prepared <- prepareLianaForCytosigComparison(liana_df)
  # Now 'liana_prepared' is ready for comparison with Cytosig results
}

}
