% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_data_wrangling.R
\name{prepareNatmiForCytosigComparison}
\alias{prepareNatmiForCytosigComparison}
\title{Prepare NATMI Data for Cytosig Comparison}
\usage{
prepareNatmiForCytosigComparison(natmi_df)
}
\arguments{
\item{natmi_df}{A dataframe resulting from NATMI analysis containing columns for ligand symbols,
receptor symbols, source and target clusters, and log2 transformed fold change of edge
expression weights.}
}
\value{
A list of dataframes, each keyed by the receiver cluster, formatted for direct comparison
with Cytosig results. This includes standardized column names (`ligand`, `receptor`, `sender_cluster`,
`receiver_cluster`, `score`) and a new `interaction` column.
}
\description{
Filters and reformats a NATMI analysis dataframe for compatibility with Cytosig comparison.
This involves filtering out infinite values from the log2 transformed fold change of edge
expression weights, renaming columns to standard ligand-receptor interaction terms, creating
an interaction identifier by concatenating ligand and receptor names, and finally splitting
the dataframe into a list of dataframes based on the receiver cluster.
}
\examples{
\dontrun{
  # Assuming 'natmi_df' is your dataframe from NATMI analysis
  natmi_prepared <- prepareNatmiForCytosigComparison(natmi_df)
  # 'natmi_prepared' is now ready for comparison with Cytosig results
}

}
