% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_data_wrangling.R
\name{preProcessConnectome}
\alias{preProcessConnectome}
\title{Preprocess Connectome Dataframe}
\usage{
preProcessConnectome(connectome_df)
}
\arguments{
\item{connectome_df}{A dataframe representing connectome data, expected to include
'source', 'target', and 'score' columns, with 'source' and 'target' representing
the sender and receiver, respectively, and 'score' indicating the interaction strength.}
}
\value{
A processed dataframe with infinite scores filtered out, columns renamed to
'sender', 'receiver', and 'prioritization_score', an additional 'interaction' column
created, and data arranged by receiver and prioritization score in descending order.
The 'prioritization_score' is scaled for comparability.
}
\description{
This function preprocesses a dataframe containing connectome data. It filters out
entries with infinite scores, renames columns for standardization, adds an
interaction column concatenating ligand and receptor names, selects relevant columns,
and arranges the dataframe based on the receiver and descending prioritization scores.
Finally, it scales the prioritization scores to make them comparable.
}
\examples{
\dontrun{
  processed_connectome_df <- preProcessConnectome(connectome_data)
  print(processed_connectome_df)
}

}
