% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_data_wrangling.R
\name{preProcessLIANA}
\alias{preProcessLIANA}
\title{Preprocess LIANA Dataframe}
\usage{
preProcessLIANA(liana_df)
}
\arguments{
\item{liana_df}{A dataframe containing ligand-receptor interaction data from LIANA,
including columns for 'ligand', 'receptor', 'source' (sender), 'target' (receiver),
and 'interaction_stat' (the statistical score representing the strength or significance
of the interaction).}
}
\value{
A processed dataframe with a new 'interaction' column, standardized column
names ('sender', 'receiver', 'prioritization_score'), and entries arranged by receiver
and prioritization score in descending order. The 'prioritization_score' is scaled
for comparability across different datasets or analysis methods.
}
\description{
Processes a dataframe resulting from LIANA (Ligand-Receptor Inference And Network
Analysis) by adding an interaction column that concatenates ligand and receptor
names, renaming columns for clarity, selecting specific columns for downstream
analysis, and arranging the entries based on the receiver and the prioritization
score in descending order. The prioritization scores are then scaled for
comparability.
}
\examples{
\dontrun{
  processed_liana_df <- preProcessLIANA(liana_data)
  print(processed_liana_df)
}

}
