% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_data_wrangling.R
\name{preProcessNATMI}
\alias{preProcessNATMI}
\title{Preprocess NATMI Dataframe}
\usage{
preProcessNATMI(natmi_df)
}
\arguments{
\item{natmi_df}{A dataframe containing NATMI analysis results. The dataframe
must include columns for ligand symbols, receptor symbols, sending and target
clusters, and log2 transformed fold change of edge expression weight.}
}
\value{
A dataframe where infinite values in the log2 transformed fold change
of edge expression weight have been removed, column names have been standardized,
an interaction column has been added, and only relevant columns are retained.
Furthermore, the prioritization score is scaled for comparability, and the dataframe
is ordered by receiver and prioritization score in descending order.
}
\description{
This function takes a dataframe containing NATMI analysis results and performs
a series of preprocessing steps including filtering, renaming, and reordering of
columns, as well as adding a new interaction column. It concludes by scaling
the prioritization scores for comparability.
}
\examples{
\dontrun{
  processed_natmi_df <- preProcessNATMI(natmi_results_all)
  print(processed_natmi_df)
}

}
