% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_data_wrangling.R
\name{pull_top_regulons_cts_delta_matrix}
\alias{pull_top_regulons_cts_delta_matrix}
\title{Create a Matrix of Regulon Delta Values Across Cell Types}
\usage{
pull_top_regulons_cts_delta_matrix(
  regulon_deltas_all_clusters,
  top_regulons_and_cts
)
}
\arguments{
\item{regulon_deltas_all_clusters}{A list where each element is a dataframe
containing the delta Pagoda scores for regulons within a specific cell type.
Each list element's name corresponds to a cell type, and each dataframe must
contain at least the columns `name` for the regulon names and `deltaPagoda`
for their scores.}

\item{top_regulons_and_cts}{A list with two elements: `top_regulons`, a vector
containing the names of the top regulons to include in the matrix, and `cts`,
a vector containing the names of the cell types to be included as columns in
the matrix.}
}
\value{
A matrix with rows corresponding to top regulons and columns to cell types.
The entries of the matrix are the delta Pagoda scores of the regulons in each cell
type, with 0s where a regulon is not present in a cell type.
}
\description{
Constructs a matrix where each row represents a top regulon and each column
represents a cell type (CT). The values in the matrix are the delta Pagoda
scores for the corresponding regulon in each CT. If a regulon is not present
in a CT, its delta Pagoda score in the matrix is set to 0.
}
\examples{
\dontrun{
  regulon_deltas <- list(
    CT1 = data.frame(name = c("Regulon1", "Regulon2"), deltaPagoda = c(2.3, -1.5)),
    CT2 = data.frame(name = c("Regulon2", "Regulon3"), deltaPagoda = c(3.1, 0.5))
  )
  top_regulons_and_cts <- list(
    top_regulons = c("Regulon1", "Regulon2", "Regulon3"),
    cts = c("CT1", "CT2")
  )
  delta_matrix <- pull_top_regulons_cts_delta_matrix(regulon_deltas, top_regulons_and_cts)
  print(delta_matrix)
}

}
