% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_utilities.R
\name{replaceNAw0}
\alias{replaceNAw0}
\title{Replace NA values with 0 in a vector}
\usage{
replaceNAw0(vector)
}
\arguments{
\item{vector}{A numeric vector that may contain NA values.}
}
\value{
A numeric vector where all NA values have been replaced with 0.
}
\description{
This function takes a vector and replaces all NA values with 0.
The modified vector with 0s replacing the NAs is then returned.
}
\examples{
sample_vector <- c(1, NA, 3, NA, 5)
replaceNAw0(sample_vector)
}
