% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_data_wrangling.R
\name{scale_prioritization_score}
\alias{scale_prioritization_score}
\title{Scale Prioritization Score}
\usage{
scale_prioritization_score(df, score_column)
}
\arguments{
\item{df}{A dataframe containing the prioritization scores to be scaled.}

\item{score_column}{A string specifying the name of the column that contains
the prioritization scores.}
}
\value{
A dataframe with an additional column `scaled_score` containing the
scaled prioritization scores.
}
\description{
This function scales the prioritization scores in a dataframe column.
If all values are non-negative, they are scaled between 0 and 1. If negative
values are present, they are scaled to a range between -1 and 1.
}
\examples{
\dontrun{
  # Assuming 'your_dataframe' has a score column named 'priority_score'
  scaled_df <- scale_prioritization_score(your_dataframe, 'priority_score')
  print(scaled_df)
}

}
