% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_data_wrangling.R
\name{summarizeZScores}
\alias{summarizeZScores}
\title{Summarize Z-Scores Across Clusters}
\usage{
summarizeZScores(z_score_files, z_score_folder, mapping_table)
}
\arguments{
\item{z_score_files}{A vector of filenames (including paths) pointing to the CSV
files containing Z-scores for ligands. Each file represents a different cluster
or sample.}

\item{z_score_folder}{folder where the CSV files containing Z-scores are located}

\item{mapping_table}{a mapping data frame to map from cytosig ligand proteins to gene names}
}
\value{
A list where each element is named after the input file (minus the '.csv'
extension) and contains a vector of median Z-scores for each ligand within that
file.
}
\description{
This function calculates the median Z-score for each ligand across different
clusters or samples, provided a list of Z-score files. Each file is expected
to contain Z-scores for all ligands in a particular cluster or sample.
}
\examples{
\dontrun{
  z_score_files <- c("/path/to/cluster1_z_scores.csv", "/path/to/cluster2_z_scores.csv")
  median_z_scores <- summarizeZScores(z_score_files)
  print(median_z_scores)
}

}
