% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{writeH5ADObjects}
\alias{writeH5ADObjects}
\title{Write H5AD Files for Each Cluster in a Seurat Object}
\usage{
writeH5ADObjects(seurat_object, pre_processing_path)
}
\arguments{
\item{seurat_object}{A Seurat object containing the data to be processed.}

\item{pre_processing_path}{A string specifying the path to the directory where the h5ad files will be saved.}
}
\value{
None. The function writes .h5ad files to the specified directory.
}
\description{
This function writes .h5ad files for each cluster in a given Seurat object.
Each cluster is processed and saved as a separate .h5ad file in a specified directory.
}
\examples{
\dontrun{
writeH5ADObjects(seurat_object, "path/to/pre_processing")
}
}
